Create API shell
