--- assembly/internal/performance.ts
@@ -1,39 +1,0 @@
-// @ts-ignore decorators are valid here
-@external("__aspect", "performanceEnabled") @global
-export declare function performanceEnabled(value: bool): void;
-
-// @ts-ignore decorators are valid here
-@external("__aspect", "maxSamples") @global
-export declare function maxSamples(value: f64): void;
-
-// @ts-ignore decorators are valid here
-@external("__aspect", "maxTestRunTime") @global
-export declare function maxTestRunTime(value: f64): void;
-
-// @ts-ignore decorators are valid here
-@external("__aspect", "roundDecimalPlaces") @global
-export declare function roundDecimalPlaces(value: i32): void;
-
-// @ts-ignore decorators are valid here
-@external("__aspect", "reportAverage") @global
-export declare function reportAverage(value: bool): void;
-
-// @ts-ignore decorators are valid here
-@external("__aspect", "reportMedian") @global
-export declare function reportMedian(value: bool): void;
-
-// @ts-ignore decorators are valid here
-@external("__aspect", "reportStdDev") @global
-export declare function reportStdDev(value: bool): void;
-
-// @ts-ignore decorators are valid here
-@external("__aspect", "reportMax") @global
-export declare function reportMax(value: bool): void;
-
-// @ts-ignore decorators are valid here
-@external("__aspect", "reportMin") @global
-export declare function reportMin(value: bool): void;
-
-// @ts-ignore decorators are valid here
-@external("__aspect", "reportVariance") @global
-export declare function reportVariance(value: bool): void;
