+++ lib/as-pect.amd.js
@@ -12,12 +12,12 @@ var __extends = (this && this.__extends) || (function () {
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
     };
 })();
+var __importDefault = (this && this.__importDefault) || function (mod) {
+    return (mod && mod.__esModule) ? mod : { "default": mod };
+};
 var __makeTemplateObject = (this && this.__makeTemplateObject) || function (cooked, raw) {
     if (Object.defineProperty) { Object.defineProperty(cooked, "raw", { value: raw }); } else { cooked.raw = raw; }
     return cooked;
-};
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 var __importStar = (this && this.__importStar) || function (mod) {
     if (mod && mod.__esModule) return mod;
@@ -27,6 +27,39 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 define("test/IWarning", ["require", "exports"], function (require, exports) {
+    "use strict";
+    Object.defineProperty(exports, "__esModule", { value: true });
+});
+define("util/IPerformanceConfiguration", ["require", "exports"], function (require, exports) {
+    "use strict";
+    Object.defineProperty(exports, "__esModule", { value: true });
+    function createDefaultPerformanceConfiguration() {
+        return {
+            /** Enable performance statistics gathering. */
+            enabled: false,
+            /** Set the maximum number of samples to run for each test. */
+            maxSamples: 10000,
+            /** Set the maximum test run time in milliseconds. */
+            maxTestRunTime: 2000,
+            /** Set the number of decimal places to round to. */
+            roundDecimalPlaces: 3,
+            /** Report the median time in the default reporter. */
+            reportMedian: true,
+            /** Report the average time in milliseconds. */
+            reportAverage: true,
+            /** Report the standard deviation. */
+            reportStandardDeviation: false,
+            /** Report the maximum run time in milliseconds. */
+            reportMax: false,
+            /** Report the minimum run time in milliseconds. */
+            reportMin: false,
+            /** Report the variance/ */
+            reportVariance: false,
+        };
+    }
+    exports.createDefaultPerformanceConfiguration = createDefaultPerformanceConfiguration;
+});
+define("reporter/IWriteable", ["require", "exports"], function (require, exports) {
     "use strict";
     Object.defineProperty(exports, "__esModule", { value: true });
 });
@@ -300,11 +333,479 @@ define("test/TestReporter", ["require", "exports"], function (require, exports) 
     }());
     exports.TestReporter = TestReporter;
 });
-define("reporter/IWriteable", ["require", "exports"], function (require, exports) {
+define("reporter/EmptyReporter", ["require", "exports", "test/TestReporter"], function (require, exports, TestReporter_1) {
+    "use strict";
+    Object.defineProperty(exports, "__esModule", { value: true });
+    var EmptyReporter = /** @class */ (function (_super) {
+        __extends(EmptyReporter, _super);
+        function EmptyReporter() {
+            return _super !== null && _super.apply(this, arguments) || this;
+        }
+        EmptyReporter.prototype.onFinish = function () { };
+        EmptyReporter.prototype.onGroupFinish = function () { };
+        EmptyReporter.prototype.onGroupStart = function () { };
+        EmptyReporter.prototype.onStart = function () { };
+        EmptyReporter.prototype.onTestFinish = function () { };
+        EmptyReporter.prototype.onTestStart = function () { };
+        EmptyReporter.prototype.onTodo = function () { };
+        return EmptyReporter;
+    }(TestReporter_1.TestReporter));
+    exports.EmptyReporter = EmptyReporter;
+});
+define("util/IAspectExports", ["require", "exports"], function (require, exports) {
     "use strict";
     Object.defineProperty(exports, "__esModule", { value: true });
 });
-define("reporter/DefaultTestReporter", ["require", "exports", "chalk", "test/TestReporter"], function (require, exports, chalk_1, TestReporter_1) {
+define("test/TestContext", ["require", "exports", "reporter/EmptyReporter", "util/ActualValue", "util/LogValue"], function (require, exports, EmptyReporter_1, ActualValue_1, LogValue_2) {
+    "use strict";
+    Object.defineProperty(exports, "__esModule", { value: true });
+    var wasmFilter = function (input) { return /wasm-function/i.test(input); };
+    var TestContext = /** @class */ (function () {
+        function TestContext(props) {
+            this.reporter = new EmptyReporter_1.EmptyReporter();
+            /* istanbul ignore next */
+            this.stdout = process ? process.stdout : null;
+            /* istanbul ignore next */
+            this.stderr = process.stderr ? process.stderr : null;
+            this.performanceConfiguration = {};
+            this.testRegex = new RegExp("");
+            this.groupRegex = new RegExp("");
+            this.fileName = "";
+            this.wasm = null;
+            this.actual = null;
+            this.expected = null;
+            this.logTarget = {
+                logs: [],
+                stack: "",
+                message: "",
+            };
+            if (props) {
+                if (props.reporter)
+                    this.reporter = props.reporter;
+                if (props.stdout)
+                    this.stdout = props.stdout;
+                if (props.stderr)
+                    this.stderr = props.stderr;
+                if (props.performanceConfiguration)
+                    this.performanceConfiguration = props.performanceConfiguration;
+                if (props.testRegex)
+                    this.testRegex = props.testRegex;
+                if (props.groupRegex)
+                    this.groupRegex = props.groupRegex;
+                if (props.fileName)
+                    this.fileName = props.fileName;
+            }
+        }
+        /**
+         * This method creates a WebAssembly imports object with all the TestContext functions
+         * bound to the TestContext.
+         *
+         * @param {any[]} imports - Every import item specified.
+         */
+        TestContext.prototype.createImports = function () {
+            var _this = this;
+            var imports = [];
+            for (var _i = 0; _i < arguments.length; _i++) {
+                imports[_i] = arguments[_i];
+            }
+            var result = Object.assign.apply(Object, [{}].concat(imports, [{
+                    __aspect: {
+                        clearExpected: this.clearExpected.bind(this),
+                        debug: this.debug.bind(this),
+                        tryCall: this.tryCall.bind(this),
+                        logNull: this.logNull.bind(this),
+                        logReference: this.logReference.bind(this),
+                        logString: this.logString.bind(this),
+                        logValue: this.logValue.bind(this),
+                        reportActualNull: this.reportActualNull.bind(this),
+                        reportExpectedNull: this.reportExpectedNull.bind(this),
+                        reportActualValue: this.reportActualValue.bind(this),
+                        reportExpectedValue: this.reportExpectedValue.bind(this),
+                        reportActualReference: this.reportActualReference.bind(this),
+                        reportExpectedReference: this.reportExpectedReference.bind(this),
+                        reportActualString: this.reportActualString.bind(this),
+                        reportExpectedString: this.reportExpectedString.bind(this),
+                        reportExpectedTruthy: this.reportExpectedTruthy.bind(this),
+                        reportExpectedFalsy: this.reportExpectedFalsy.bind(this),
+                        reportExpectedFinite: this.reportExpectedFinite.bind(this),
+                    },
+                }]));
+            result.env = result.env || {};
+            var previousAbort = (result.env.abort) || (function () { });
+            result.env.abort = function () {
+                var args = [];
+                for (var _i = 0; _i < arguments.length; _i++) {
+                    args[_i] = arguments[_i];
+                }
+                previousAbort.apply(void 0, args);
+                // @ts-ignore
+                _this.abort.apply(_this, args);
+            };
+            return result;
+        };
+        /**
+         * This function reports an actual null value.
+         */
+        TestContext.prototype.reportActualNull = function () {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "null";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.value = null;
+            this.actual = value;
+        };
+        /**
+         * This function reports an expected null value.
+         *
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
+         */
+        TestContext.prototype.reportExpectedNull = function (negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "null";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.negated = negated === 1;
+            value.value = null;
+            this.expected = value;
+        };
+        /**
+         * This function reports an actual numeric value.
+         *
+         * @param {number} numericValue - The value to be expected.
+         */
+        TestContext.prototype.reportActualValue = function (numericValue) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = numericValue.toString();
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.value = numericValue;
+            this.actual = value;
+        };
+        /**
+         * This function reports an expected numeric value.
+         *
+         * @param {number} numericValue - The value to be expected
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
+         */
+        TestContext.prototype.reportExpectedValue = function (numericValue, negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = numericValue.toString();
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.negated = negated === 1;
+            value.value = numericValue;
+            this.expected = value;
+        };
+        /**
+         * This function reports an actual reference value.
+         *
+         * @param {number} referencePointer - The actual reference pointer.
+         * @param {number} offset - The size of the reference in bytes.
+         */
+        TestContext.prototype.reportActualReference = function (referencePointer, offset) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "Reference Value";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.pointer = referencePointer;
+            value.offset = offset;
+            value.bytes = Array.from(this.wasm.U8.slice(referencePointer, referencePointer + offset));
+            value.value = referencePointer;
+            this.actual = value;
+        };
+        /**
+         * This function reports an expected reference value.
+         *
+         * @param {number} referencePointer - The expected reference pointer.
+         * @param {number} offset - The size of the reference in bytes.
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
+         */
+        TestContext.prototype.reportExpectedReference = function (referencePointer, offset, negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "Reference Value";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.pointer = referencePointer;
+            value.offset = offset;
+            value.bytes = Array.from(this.wasm.U8.slice(referencePointer, referencePointer + offset));
+            value.negated = negated === 1;
+            value.value = referencePointer;
+            this.expected = value;
+        };
+        /**
+         * This function reports an expected truthy value.
+         *
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
+         */
+        TestContext.prototype.reportExpectedTruthy = function (negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "Truthy Value";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.negated = negated === 1;
+            this.expected = value;
+        };
+        /**
+         * This function reports an expected falsy value.
+         *
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
+         */
+        TestContext.prototype.reportExpectedFalsy = function (negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "Falsy Value";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.negated = negated === 1;
+            this.expected = value;
+        };
+        /**
+         * This function reports an expected finite value.
+         *
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
+         */
+        TestContext.prototype.reportExpectedFinite = function (negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "Finite Value";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.negated = negated === 1;
+            this.expected = value;
+        };
+        /**
+         * This function reports an actual string value.
+         *
+         * @param {number} stringPointer - A pointer that points to the actual string.
+         */
+        TestContext.prototype.reportActualString = function (stringPointer) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = this.wasm.__getString(stringPointer);
+            value.pointer = stringPointer;
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.value = stringPointer;
+            this.actual = value;
+        };
+        /**
+         * This function reports an expected string value.
+         *
+         * @param {number} stringPointer - A pointer that points to the expected string.
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
+         */
+        TestContext.prototype.reportExpectedString = function (stringPointer, negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = this.wasm.__getString(stringPointer);
+            value.pointer = stringPointer;
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.negated = negated === 1;
+            value.value = stringPointer;
+            this.expected = value;
+        };
+        /**
+         * This function overrides the provided AssemblyScript `env.abort()` function to catch abort
+         * reasons.
+         *
+         * @param {number} reasonPointer - This points to the message value that causes the expectation to
+         * fail.
+         * @param {number} _fileNamePointer - The file name that reported the error. (Ignored)
+         * @param {number} _line - The line that reported the error. (Ignored)
+         * @param {number} _col - The column that reported the error. (Ignored)
+         */
+        TestContext.prototype.abort = function (reasonPointer, _fileNamePointer, _line, _col) {
+            this.logTarget.message = this.wasm.__getString(reasonPointer);
+        };
+        /**
+         * Gets a log stack trace.
+         */
+        TestContext.prototype.getLogStackTrace = function () {
+            return new Error("Get stack trace.")
+                .stack
+                .toString()
+                .split("\n")
+                .slice(1)
+                .filter(wasmFilter)
+                .join("\n");
+        };
+        /**
+         * Gets an error stack trace.
+         */
+        TestContext.prototype.getErrorStackTrace = function (ex) {
+            var stackItems = ex.stack.toString().split("\n");
+            return [stackItems[0]].concat(stackItems.slice(1).filter(wasmFilter)).join("\n");
+        };
+        /**
+         * This is called to stop the debugger.  e.g. `node --inspect-brk asp`.
+         */
+        TestContext.prototype.debug = function () { debugger; };
+        /**
+         * This is a web assembly utility function that wraps a function call in a try catch block to
+         * report success or failure.
+         *
+         * @param {number} pointer - The function pointer to call. It must accept no parameters and return
+         * void.
+         * @returns {1 | 0} - If the callback was run successfully without error, it returns 1, else it
+         * returns 0.
+         */
+        TestContext.prototype.tryCall = function (pointer) {
+            if (pointer === -1)
+                return 1;
+            try {
+                this.wasm.__call(pointer);
+            }
+            catch (ex) {
+                this.logTarget.stack = this.getErrorStackTrace(ex);
+                return 0;
+            }
+            return 1;
+        };
+        /**
+         * Log a null value to the reporter.
+         */
+        TestContext.prototype.logNull = function () {
+            // create a new log value
+            var value = new LogValue_2.LogValue();
+            var target = this.logTarget;
+            // collect log metadata
+            value.stack = this.getLogStackTrace();
+            value.message = "null";
+            value.target = target;
+            // push the log value to the logs
+            target.logs.push(value);
+        };
+        /**
+         * This function is called after each expectation if the expectation passes. This prevents other
+         * unreachable() conditions that throw errors to report actual and expected values too.
+         */
+        TestContext.prototype.clearExpected = function () {
+            this.expected = null;
+            this.actual = null;
+            this.logTarget.stack = "";
+        };
+        /**
+         * Log a reference to the reporter.
+         *
+         * @param {number} referencePointer - The pointer to the reference.
+         * @param {number} offset - The offset of the reference.
+         */
+        TestContext.prototype.logReference = function (referencePointer, offset) {
+            var value = new LogValue_2.LogValue();
+            var target = this.logTarget;
+            value.bytes = Array.from(this.wasm.U8.slice(referencePointer, referencePointer + offset));
+            value.message = "Reference Type";
+            value.offset = offset;
+            value.pointer = referencePointer;
+            value.stack = this.getLogStackTrace();
+            value.target = target;
+            value.value = referencePointer;
+            // push the log value to the logs
+            target.logs.push(value);
+        };
+        /**
+         * This adds a logged string to the current test.
+         *
+         * @param {number} pointer - The pointer to the logged string reference.
+         */
+        TestContext.prototype.logString = function (pointer) {
+            var value = new LogValue_2.LogValue();
+            var target = this.logTarget;
+            value.message = this.wasm.__getString(pointer);
+            value.offset = 0;
+            value.pointer = pointer;
+            value.stack = this.getLogStackTrace();
+            value.target = target;
+            value.value = pointer;
+            // push the log value to the logs
+            target.logs.push(value);
+        };
+        /**
+         * Log a numevalueric value to the reporter.
+         *
+         * @param {number} value - The value to be logged.
+         */
+        TestContext.prototype.logValue = function (numericValue) {
+            var value = new LogValue_2.LogValue();
+            var target = this.logTarget;
+            value.stack = this.getLogStackTrace();
+            value.message = "Value " + numericValue.toString();
+            value.value = numericValue;
+            value.target = target;
+            // push the log value to the logs
+            target.logs.push(value);
+        };
+        return TestContext;
+    }());
+    exports.TestContext = TestContext;
+});
+define("reporter/CSVTestReporter", ["require", "exports", "test/TestReporter", "csv-stringify", "fs", "path"], function (require, exports, TestReporter_2, csv_stringify_1, fs_1, path_1) {
+    "use strict";
+    Object.defineProperty(exports, "__esModule", { value: true });
+    csv_stringify_1 = __importDefault(csv_stringify_1);
+    var columns = ["Group", "Name", "Ran", "Pass", "Runtime", "Message", "Actual", "Expected", "Average", "Mean", "Max", "Min", "StdDev", "Variance"];
+    var CSVTestReporter = /** @class */ (function (_super) {
+        __extends(CSVTestReporter, _super);
+        function CSVTestReporter() {
+            var _this = _super !== null && _super.apply(this, arguments) || this;
+            _this.output = null;
+            _this.fileName = null;
+            return _this;
+        }
+        CSVTestReporter.prototype.onStart = function (suite) {
+            this.output = csv_stringify_1.default({ columns: columns });
+            var extension = path_1.extname(suite.fileName);
+            var dir = path_1.dirname(suite.fileName);
+            var base = path_1.basename(suite.fileName, extension);
+            var outPath = path_1.join(process.cwd(), dir, base + ".csv");
+            this.fileName = fs_1.createWriteStream(outPath, "utf8");
+            this.output.pipe(this.fileName);
+            this.output.write(columns);
+        };
+        CSVTestReporter.prototype.onGroupStart = function () { };
+        CSVTestReporter.prototype.onGroupFinish = function () { };
+        CSVTestReporter.prototype.onFinish = function () {
+            this.output.end();
+        };
+        CSVTestReporter.prototype.onTestStart = function () { };
+        CSVTestReporter.prototype.onTestFinish = function (group, result) {
+            this.output.write([
+                group.name,
+                result.ran ? "RAN" : "NOT RUN",
+                result.name,
+                result.pass ? "PASS" : "FAIL",
+                result.runTime.toString(),
+                result.message,
+                result.actual ? result.actual.message : "",
+                result.expected ? result.expected.message : "",
+                result.hasAverage ? result.average.toString() : "",
+                result.hasMedian ? result.median.toString() : "",
+                result.hasMax ? result.max.toString() : "",
+                result.hasMin ? result.min.toString() : "",
+                result.hasStdDev ? result.stdDev.toString() : "",
+                result.hasVariance ? result.variance.toString() : ""
+            ]);
+        };
+        CSVTestReporter.prototype.onTodo = function (group, desc) {
+            this.output.write([
+                group.name,
+                "TODO",
+                desc,
+                "",
+                "",
+                "",
+                "",
+                "",
+                "",
+                "",
+                "",
+                "",
+                "",
+                "",
+            ]);
+        };
+        return CSVTestReporter;
+    }(TestReporter_2.TestReporter));
+    exports.CSVTestReporter = CSVTestReporter;
+});
+define("reporter/DefaultTestReporter", ["require", "exports", "chalk", "test/TestReporter"], function (require, exports, chalk_1, TestReporter_3) {
     "use strict";
     Object.defineProperty(exports, "__esModule", { value: true });
     chalk_1 = __importDefault(chalk_1);
@@ -491,1146 +992,9 @@ define("reporter/DefaultTestReporter", ["require", "exports", "chalk", "test/Tes
             this.stdout.write(chalk_1.default(templateObject_25 || (templateObject_25 = __makeTemplateObject(["        {yellow ", "}\n\n"], ["        {yellow ", "}\\n\\n"])), logValue.stack.split("\n").join("\n        ")));
         };
         return DefaultTestReporter;
-    }(TestReporter_1.TestReporter));
+    }(TestReporter_3.TestReporter));
     exports.DefaultTestReporter = DefaultTestReporter;
     var templateObject_1, templateObject_2, templateObject_3, templateObject_4, templateObject_5, templateObject_6, templateObject_7, templateObject_8, templateObject_9, templateObject_10, templateObject_11, templateObject_12, templateObject_13, templateObject_14, templateObject_15, templateObject_16, templateObject_17, templateObject_18, templateObject_19, templateObject_20, templateObject_21, templateObject_22, templateObject_23, templateObject_24, templateObject_25;
-});
-define("util/timeDifference", ["require", "exports"], function (require, exports) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    exports.timeDifference = function (end, start) { return Math.round((end - start) * 1000) / 1000; };
-});
-define("util/IAspectExports", ["require", "exports"], function (require, exports) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-});
-define("util/IPerformanceConfiguration", ["require", "exports"], function (require, exports) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    function createDefaultPerformanceConfiguration() {
-        return {
-            /** Enable performance statistics gathering. */
-            enabled: false,
-            /** Set the maximum number of samples to run for each test. */
-            maxSamples: 10000,
-            /** Set the maximum test run time in milliseconds. */
-            maxTestRunTime: 2000,
-            /** Set the number of decimal places to round to. */
-            roundDecimalPlaces: 3,
-            /** Report the median time in the default reporter. */
-            reportMedian: true,
-            /** Report the average time in milliseconds. */
-            reportAverage: true,
-            /** Report the standard deviation. */
-            reportStandardDeviation: false,
-            /** Report the maximum run time in milliseconds. */
-            reportMax: false,
-            /** Report the minimum run time in milliseconds. */
-            reportMin: false,
-            /** Report the variance/ */
-            reportVariance: false,
-        };
-    }
-    exports.createDefaultPerformanceConfiguration = createDefaultPerformanceConfiguration;
-});
-define("test/TestCollector", ["require", "exports", "util/LogValue", "util/ActualValue", "test/TestGroup", "util/IPerformanceConfiguration", "test/TestResult"], function (require, exports, LogValue_2, ActualValue_1, TestGroup_1, IPerformanceConfiguration_1, TestResult_1) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    var wasmFilter = function (input) { return /wasm-function/i.test(input); };
-    /**
-     * This class is responsible for collecting all the tests in a test binary.
-     */
-    var TestCollector = /** @class */ (function () {
-        function TestCollector(props) {
-            this.wasm = null;
-            // test group values
-            this.groupStack = [new TestGroup_1.TestGroup()];
-            this.testGroups = [];
-            this.logTarget = this.groupStack[0];
-            // public warning/error lists
-            this.errors = [];
-            // public fileName
-            this.fileName = "";
-            // test state machine values
-            this.stack = "";
-            this.message = "";
-            this.actual = null;
-            this.expected = null;
-            this.performanceConfiguration = IPerformanceConfiguration_1.createDefaultPerformanceConfiguration();
-            /**
-             * This value is used to detect if an `expect()` function call was used outside of a test
-             * function. If a reportExpected or reportActual function is called before the `context.run()`
-             * method is called, it should prevent the `run()` method from running the tests and report a
-             * failure.
-             */
-            this.ready = false;
-            /**
-             * These are the test and group filters for the binary. They must be provided before collection
-             * begins.
-             */
-            this.testRegex = new RegExp("");
-            this.groupRegex = new RegExp("");
-            if (props) {
-                /* istanbul ignore next */
-                if (props.fileName)
-                    this.fileName = props.fileName;
-                /* istanbul ignore next */
-                if (props.testRegex)
-                    this.testRegex = props.testRegex;
-                /* istanbul ignore next */
-                if (props.groupRegex)
-                    this.groupRegex = props.groupRegex;
-                /* istanbul ignore next */
-                if (props.performanceConfiguration)
-                    this.performanceConfiguration = props.performanceConfiguration;
-            }
-        }
-        /**
-         * Call this method to start the `__main()` method provided by the `as-pect` exports to start the
-         * process of test collection.
-         */
-        TestCollector.prototype.collectTests = function () {
-            // reset the performance values first, then collect the tests by calling `__main()`
-            this.resetPerformanceValues();
-            this.wasm.__main();
-        };
-        /**
-         * This method creates a WebAssembly imports object with all the TestContext functions
-         * bound to the TestContext.
-         *
-         * @param {any[]} imports - Every import item specified.
-         */
-        TestCollector.prototype.createImports = function () {
-            var _this = this;
-            var imports = [];
-            for (var _i = 0; _i < arguments.length; _i++) {
-                imports[_i] = arguments[_i];
-            }
-            var result = Object.assign.apply(Object, [{}].concat(imports, [{
-                    __aspect: {
-                        clearExpected: this.clearExpected.bind(this),
-                        debug: this.debug.bind(this),
-                        tryCall: this.tryCall.bind(this),
-                        logNull: this.logNull.bind(this),
-                        logReference: this.logReference.bind(this),
-                        logString: this.logString.bind(this),
-                        logValue: this.logValue.bind(this),
-                        reportDescribe: this.reportDescribe.bind(this),
-                        reportEndDescribe: this.reportEndDescribe.bind(this),
-                        reportTest: this.reportTest.bind(this),
-                        reportBeforeEach: this.reportBeforeEach.bind(this),
-                        reportBeforeAll: this.reportBeforeAll.bind(this),
-                        reportAfterEach: this.reportAfterEach.bind(this),
-                        reportAfterAll: this.reportAfterAll.bind(this),
-                        reportTodo: this.reportTodo.bind(this),
-                        reportActualNull: this.reportActualNull.bind(this),
-                        reportExpectedNull: this.reportExpectedNull.bind(this),
-                        reportActualValue: this.reportActualValue.bind(this),
-                        reportExpectedValue: this.reportExpectedValue.bind(this),
-                        reportActualReference: this.reportActualReference.bind(this),
-                        reportExpectedReference: this.reportExpectedReference.bind(this),
-                        reportActualString: this.reportActualString.bind(this),
-                        reportExpectedString: this.reportExpectedString.bind(this),
-                        reportExpectedTruthy: this.reportExpectedTruthy.bind(this),
-                        reportExpectedFalsy: this.reportExpectedFalsy.bind(this),
-                        reportExpectedFinite: this.reportExpectedFinite.bind(this),
-                        reportNegatedTest: this.reportNegatedTest.bind(this),
-                        performanceEnabled: this.performanceEnabled.bind(this),
-                        maxSamples: this.maxSamples.bind(this),
-                        maxTestRunTime: this.maxTestRunTime.bind(this),
-                        roundDecimalPlaces: this.roundDecimalPlaces.bind(this),
-                        reportAverage: this.reportAverage.bind(this),
-                        reportMedian: this.reportMedian.bind(this),
-                        reportStdDev: this.reportStdDev.bind(this),
-                        reportMax: this.reportMax.bind(this),
-                        reportMin: this.reportMin.bind(this),
-                        reportVariance: this.reportVariance.bind(this),
-                    },
-                }]));
-            result.env = result.env || {};
-            var previousAbort = (result.env.abort) || (function () { });
-            result.env.abort = function () {
-                var args = [];
-                for (var _i = 0; _i < arguments.length; _i++) {
-                    args[_i] = arguments[_i];
-                }
-                previousAbort.apply(void 0, args);
-                // @ts-ignore
-                _this.abort.apply(_this, args);
-            };
-            return result;
-        };
-        /**
-         * This is called to stop the debugger.  e.g. `node --inspect-brk asp`.
-         */
-        TestCollector.prototype.debug = function () { debugger; };
-        /**
-         * This is a web assembly utility function that wraps a function call in a try catch block to
-         * report success or failure.
-         *
-         * @param {number} pointer - The function pointer to call. It must accept no parameters and return
-         * void.
-         * @returns {1 | 0} - If the callback was run successfully without error, it returns 1, else it
-         * returns 0.
-         */
-        TestCollector.prototype.tryCall = function (pointer) {
-            if (pointer === -1)
-                return 1;
-            try {
-                this.wasm.__call(pointer);
-            }
-            catch (ex) {
-                this.stack = this.getErrorStackTrace(ex);
-                return 0;
-            }
-            return 1;
-        };
-        /**
-         * Log a null value to the reporter.
-         */
-        TestCollector.prototype.logNull = function () {
-            // create a new log value
-            var value = new LogValue_2.LogValue();
-            var target = this.logTarget;
-            // collect log metadata
-            value.stack = this.getLogStackTrace();
-            value.message = "null";
-            value.target = target;
-            // push the log value to the logs
-            target.logs.push(value);
-        };
-        /**
-         * This function is called after each expectation if the expectation passes. This prevents other
-         * unreachable() conditions that throw errors to report actual and expected values too.
-         */
-        TestCollector.prototype.clearExpected = function () {
-            this.expected = null;
-            this.actual = null;
-            this.stack = "";
-        };
-        /**
-         * Log a reference to the reporter.
-         *
-         * @param {number} referencePointer - The pointer to the reference.
-         * @param {number} offset - The offset of the reference.
-         */
-        TestCollector.prototype.logReference = function (referencePointer, offset) {
-            var value = new LogValue_2.LogValue();
-            var target = this.logTarget;
-            value.bytes = Array.from(this.wasm.U8.slice(referencePointer, referencePointer + offset));
-            value.message = "Reference Type";
-            value.offset = offset;
-            value.pointer = referencePointer;
-            value.stack = this.getLogStackTrace();
-            value.target = target;
-            value.value = referencePointer;
-            // push the log value to the logs
-            target.logs.push(value);
-        };
-        /**
-         * This adds a logged string to the current test.
-         *
-         * @param {number} pointer - The pointer to the logged string reference.
-         */
-        TestCollector.prototype.logString = function (pointer) {
-            var value = new LogValue_2.LogValue();
-            var target = this.logTarget;
-            value.message = this.wasm.__getString(pointer);
-            value.offset = 0;
-            value.pointer = pointer;
-            value.stack = this.getLogStackTrace();
-            value.target = target;
-            value.value = pointer;
-            // push the log value to the logs
-            target.logs.push(value);
-        };
-        /**
-         * Log a numevalueric value to the reporter.
-         *
-         * @param {number} value - The value to be logged.
-         */
-        TestCollector.prototype.logValue = function (numericValue) {
-            var value = new LogValue_2.LogValue();
-            var target = this.logTarget;
-            value.stack = this.getLogStackTrace();
-            value.message = "Value " + numericValue.toString();
-            value.value = numericValue;
-            value.target = target;
-            // push the log value to the logs
-            target.logs.push(value);
-        };
-        /**
-         * This web assembly linked function creates a test group. It's called when the test suite calls
-         * the describe("test", callback) function from within AssemblyScript. It receives a pointer to
-         * the description of the tests, forks the top level test group, pushes the suiteName to a list,
-         * then pushes the forked group to the top of the test context stack.
-         *
-         * @param {number} suiteNamePointer
-         */
-        TestCollector.prototype.reportDescribe = function (suiteNamePointer) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            var nextGroup = group.fork();
-            nextGroup.name = group.name + this.wasm.__getString(suiteNamePointer);
-            nextGroup.willRun = this.groupRegex.test(nextGroup.name);
-            this.groupStack.push(nextGroup);
-            this.logTarget = nextGroup;
-        };
-        /**
-         * This web assembly linked function finishes a test group. It's called when the test suite calls
-         * the describe("test", callback) function from within AssemblyScript. It pops the current
-         * test group from the test context stack and pushes it to the final test group list.
-         */
-        TestCollector.prototype.reportEndDescribe = function () {
-            var next = this.groupStack.pop();
-            var testCount = next.tests.length;
-            /**
-             * If a describe finishes first, it happens BEFORE other describes. This means
-             * inner describe blocks run at lower priority than outer describe blocks. It also should be
-             * pushed if it willRun.
-             */
-            if (next.willRun && testCount > 0) {
-                this.testGroups.unshift(next);
-            }
-            this.logTarget = this.groupStack[this.groupStack.length - 1];
-        };
-        /**
-         * This web assembly linked function sets the group's "beforeEach" callback pointer to
-         * the current groupStackItem.
-         *
-         * @param {number} callbackPointer - The callback that should run before each test.
-         */
-        TestCollector.prototype.reportBeforeEach = function (callbackPointer) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            ;
-            group.beforeEachPointers.push(callbackPointer);
-        };
-        /**
-         * This web assembly linked function adds the group's "beforeAll" callback pointer to
-         * the current groupStackItem.
-         *
-         * @param {number} callbackPointer - The callback that should run before each test in the
-         * current context.
-         */
-        TestCollector.prototype.reportBeforeAll = function (callbackPointer) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            ;
-            group.beforeAllPointers.push(callbackPointer);
-        };
-        /**
-         * This web assembly linked function sets the group's "afterEach" callback pointer.
-         *
-         * @param {number} callbackPointer - The callback that should run before each test group.
-         */
-        TestCollector.prototype.reportAfterEach = function (callbackPointer) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            ;
-            group.afterEachPointers.push(callbackPointer);
-        };
-        /**
-         * This web assembly linked function adds the group's "afterAll" callback pointer to
-         * the current groupStackItem.
-         *
-         * @param {number} callbackPointer - The callback that should run before each test in the
-         * current context.
-         */
-        TestCollector.prototype.reportAfterAll = function (callbackPointer) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            ;
-            group.afterAllPointers.push(callbackPointer);
-        };
-        /**
-         * This web assembly linked function creates a test from the callback and the testNamePointer in
-         * the current group. It assumes that the group has already been created with the describe
-         * function. It is called when `it("description", callback)` or `test("description", callback)`
-         * is called.
-         *
-         * @param {number} testNamePointer - The test's name pointer.
-         * @param {number} callback - The test's function.
-         */
-        TestCollector.prototype.reportTest = function (testNamePointer, callback) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            if (!group.willRun)
-                return;
-            var name = this.wasm.__getString(testNamePointer);
-            if (!this.testRegex.test(name))
-                return;
-            var test = new TestResult_1.TestResult();
-            test.functionPointer = callback;
-            test.name = name;
-            test.message = "";
-            test.negated = false;
-            test.performance = this.performanceEnabledValue || false;
-            if (test.performance) {
-                test.maxSamples = !isFinite(this.maxSamplesValue)
-                    ? 10000 /* MaxSamples */
-                    : Math.min(this.maxSamplesValue, 10000 /* MaxSamples */);
-                test.maxRuntime = !isFinite(this.maxTestRunTimeValue)
-                    ? 5000 /* MaxTestRuntime */
-                    : Math.min(this.maxTestRunTimeValue, 5000 /* MaxTestRuntime */);
-                test.decimalPlaces = !isFinite(this.roundDecimalPlacesValue)
-                    ? 3
-                    : Math.max(Math.round(this.roundDecimalPlacesValue), 0 /* MinimumDecimalPlaces */);
-                test.calculateAverageValue = this.recordAverageValue || false;
-                test.calculateMedianValue = this.recordMedianValue || false;
-                test.calculateStandardDeviationValue = this.recordStdDevValue || false;
-                test.calculateMaxValue = this.recordMaxValue || false;
-                test.calculateMinValue = this.recordMinValue || false;
-                test.calculateVarianceValue = this.recordVariance || false;
-            }
-            group.tests.push(test);
-            this.resetPerformanceValues();
-        };
-        /**
-         * This web assembly linked function is responsible for reporting tests that are expected
-         * to fail. This is useful for verifying that specific application states will throw.
-         *
-         * @param {number} testNamePointer - The test's name pointer.
-         * @param {number} callback - The test's function.
-         * @param {number} message - The message associated with this test if it does not throw.
-         */
-        TestCollector.prototype.reportNegatedTest = function (testNamePointer, callback, message) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            if (!group.willRun)
-                return;
-            var name = this.wasm.__getString(testNamePointer);
-            if (!this.testRegex.test(name))
-                return;
-            var test = new TestResult_1.TestResult();
-            test.functionPointer = callback;
-            test.name = name;
-            test.message = this.wasm.__getString(message);
-            test.negated = true;
-            test.performance = this.performanceEnabledValue || false;
-            /* istanbul ignore next */
-            if (test.performance) {
-                /* istanbul ignore next */
-                test.maxSamples = !isFinite(this.maxSamplesValue)
-                    ? 10000 /* MaxSamples */
-                    : Math.min(this.maxSamplesValue, 10000 /* MaxSamples */);
-                /* istanbul ignore next */
-                test.maxRuntime = !isFinite(this.maxTestRunTimeValue)
-                    ? 5000 /* MaxTestRuntime */
-                    : Math.min(this.maxTestRunTimeValue, 5000 /* MaxTestRuntime */);
-                /* istanbul ignore next */
-                test.decimalPlaces = !isFinite(this.roundDecimalPlacesValue)
-                    ? 3
-                    : Math.max(Math.round(this.roundDecimalPlacesValue), 0 /* MinimumDecimalPlaces */);
-                /* istanbul ignore next */
-                test.calculateAverageValue = this.recordAverageValue || false;
-                /* istanbul ignore next */
-                test.calculateMedianValue = this.recordMedianValue || false;
-                /* istanbul ignore next */
-                test.calculateStandardDeviationValue = this.recordStdDevValue || false;
-                /* istanbul ignore next */
-                test.calculateMaxValue = this.recordMaxValue || false;
-                /* istanbul ignore next */
-                test.calculateMinValue = this.recordMaxValue || false;
-                /* istanbul ignore next */
-                test.calculateVarianceValue = this.recordVariance || false;
-            }
-            group.tests.push(test);
-            this.resetPerformanceValues();
-        };
-        /**
-         * This function reports a single "todo" item in a test suite.
-         *
-         * @param {number} todoPointer - The todo description string pointer.
-         */
-        TestCollector.prototype.reportTodo = function (todoPointer) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            group.todos.push(this.wasm.__getString(todoPointer));
-        };
-        /**
-         * This function reports an actual null value.
-         */
-        TestCollector.prototype.reportActualNull = function () {
-            if (!this.ready) {
-                this.reportInvalidExpectCall();
-                return;
-            }
-            var value = new ActualValue_1.ActualValue();
-            value.message = "null";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.value = null;
-            this.actual = value;
-        };
-        /**
-         * This function reports an expected null value.
-         *
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedNull = function (negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = "null";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.negated = negated === 1;
-            value.value = null;
-            this.expected = value;
-        };
-        /**
-         * This function reports an actual numeric value.
-         *
-         * @param {number} numericValue - The value to be expected.
-         */
-        TestCollector.prototype.reportActualValue = function (numericValue) {
-            if (!this.ready) {
-                this.reportInvalidExpectCall();
-                return;
-            }
-            var value = new ActualValue_1.ActualValue();
-            value.message = numericValue.toString();
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.value = numericValue;
-            this.actual = value;
-        };
-        /**
-         * This function reports an expected numeric value.
-         *
-         * @param {number} numericValue - The value to be expected
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedValue = function (numericValue, negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = numericValue.toString();
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.negated = negated === 1;
-            value.value = numericValue;
-            this.expected = value;
-        };
-        /**
-         * This function reports an actual reference value.
-         *
-         * @param {number} referencePointer - The actual reference pointer.
-         * @param {number} offset - The size of the reference in bytes.
-         */
-        TestCollector.prototype.reportActualReference = function (referencePointer, offset) {
-            if (!this.ready) {
-                this.reportInvalidExpectCall();
-                return;
-            }
-            var value = new ActualValue_1.ActualValue();
-            value.message = "Reference Value";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.pointer = referencePointer;
-            value.offset = offset;
-            value.bytes = Array.from(this.wasm.U8.slice(referencePointer, referencePointer + offset));
-            value.value = referencePointer;
-            this.actual = value;
-        };
-        /**
-         * This function reports an expected reference value.
-         *
-         * @param {number} referencePointer - The expected reference pointer.
-         * @param {number} offset - The size of the reference in bytes.
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedReference = function (referencePointer, offset, negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = "Reference Value";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.pointer = referencePointer;
-            value.offset = offset;
-            value.bytes = Array.from(this.wasm.U8.slice(referencePointer, referencePointer + offset));
-            value.negated = negated === 1;
-            value.value = referencePointer;
-            this.expected = value;
-        };
-        /**
-         * This function reports an expected truthy value.
-         *
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedTruthy = function (negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = "Truthy Value";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.negated = negated === 1;
-            this.expected = value;
-        };
-        /**
-         * This function reports an expected falsy value.
-         *
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedFalsy = function (negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = "Falsy Value";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.negated = negated === 1;
-            this.expected = value;
-        };
-        /**
-         * This function reports an expected finite value.
-         *
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedFinite = function (negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = "Finite Value";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.negated = negated === 1;
-            this.expected = value;
-        };
-        /**
-         * This function reports an actual string value.
-         *
-         * @param {number} stringPointer - A pointer that points to the actual string.
-         */
-        TestCollector.prototype.reportActualString = function (stringPointer) {
-            if (!this.ready) {
-                this.reportInvalidExpectCall();
-                return;
-            }
-            var value = new ActualValue_1.ActualValue();
-            value.message = this.wasm.__getString(stringPointer);
-            value.pointer = stringPointer;
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.value = stringPointer;
-            this.actual = value;
-        };
-        /**
-         * This function reports an expected string value.
-         *
-         * @param {number} stringPointer - A pointer that points to the expected string.
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedString = function (stringPointer, negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = this.wasm.__getString(stringPointer);
-            value.pointer = stringPointer;
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.negated = negated === 1;
-            value.value = stringPointer;
-            this.expected = value;
-        };
-        /**
-         * This function overrides the provided AssemblyScript `env.abort()` function to catch abort
-         * reasons.
-         *
-         * @param {number} reasonPointer - This points to the message value that causes the expectation to
-         * fail.
-         * @param {number} _fileNamePointer - The file name that reported the error. (Ignored)
-         * @param {number} _line - The line that reported the error. (Ignored)
-         * @param {number} _col - The column that reported the error. (Ignored)
-         */
-        TestCollector.prototype.abort = function (reasonPointer, _fileNamePointer, _line, _col) {
-            this.message = this.wasm.__getString(reasonPointer);
-        };
-        /**
-         * This web assembly linked function modifies the state machine to enable
-         * performance for the following test.
-         *
-         * @param {1 | 0} value - A value indicating if performance should be enabled.
-         */
-        TestCollector.prototype.performanceEnabled = function (value) {
-            this.performanceEnabledValue = value === 1;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to set the maximum number of
-         * samples for the following test.
-         *
-         * @param {number} value - The maximum number of samples to collect for the following test.
-         */
-        TestCollector.prototype.maxSamples = function (value) {
-            this.maxSamplesValue = value;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to set the maximum amount of
-         * time to run the following test in milliseconds
-         *
-         * @param {number} value - The maximum number of milliseconds to run the following test.
-         */
-        TestCollector.prototype.maxTestRunTime = function (value) {
-            this.maxTestRunTimeValue = value;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to set the number of decimal places
-         * to round all the statistics to.
-         *
-         * @param {number} value - The number of decimal places to round to.
-         */
-        TestCollector.prototype.roundDecimalPlaces = function (value) {
-            this.roundDecimalPlacesValue = value;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to cause the next test to report
-         * an average run time.
-         *
-         * @param {1 | 0} value - A boolean indicating if the average should be reported.
-         */
-        TestCollector.prototype.reportAverage = function (value) {
-            this.recordAverageValue = value === 1;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to cause the next test to report
-         * an median run time.
-         *
-         * @param {1 | 0} value - A boolean indicating if the median should be reported.
-         */
-        TestCollector.prototype.reportMedian = function (value) {
-            this.recordMedianValue = value === 1;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to cause the next test to report
-         * a standard deviation calculation on the run times.
-         *
-         * @param {1 | 0} value - A boolean indicating if the standard deviation should be reported.
-         */
-        TestCollector.prototype.reportStdDev = function (value) {
-            this.recordStdDevValue = value === 1;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to cause the next test to report
-         * the maximum run time for this test.
-         *
-         * @param {1 | 0} value - A boolean indicating if the max should be reported.
-         */
-        TestCollector.prototype.reportMax = function (value) {
-            this.recordMaxValue = value === 1;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to cause the next test to report
-         * the minimum run time for this test.
-         *
-         * @param {1 | 0} value - A boolean indicating if the min should be reported.
-         */
-        TestCollector.prototype.reportMin = function (value) {
-            this.recordMinValue = value === 1;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to cause the next test to report
-         * the variance of the run times for this test.
-         *
-         * @param {1 | 0} value - A boolean indicating if the min should be reported.
-         */
-        TestCollector.prototype.reportVariance = function (value) {
-            this.recordVariance = value === 1;
-        };
-        /**
-         * This method reports to the TestContext that an expect function call was used outside of the
-         * intended test functions.
-         */
-        TestCollector.prototype.reportInvalidExpectCall = function () {
-            this.errors.push({
-                type: "InvalidExpectCall",
-                message: "An expect() function call was used outside of a test function in " + this.fileName + ".",
-                stackTrace: this.getLogStackTrace(),
-            });
-        };
-        /**
-         * Gets an error stack trace.
-         */
-        TestCollector.prototype.getErrorStackTrace = function (ex) {
-            var stackItems = ex.stack.toString().split("\n");
-            return [stackItems[0]].concat(stackItems.slice(1).filter(wasmFilter)).join("\n");
-        };
-        /**
-         * Reset all the performance values to the configured values.
-         */
-        TestCollector.prototype.resetPerformanceValues = function () {
-            this.performanceEnabledValue = this.performanceConfiguration.enabled;
-            this.maxSamplesValue = this.performanceConfiguration.maxSamples;
-            this.maxTestRunTimeValue = this.performanceConfiguration.maxTestRunTime;
-            this.roundDecimalPlacesValue = this.performanceConfiguration.roundDecimalPlaces;
-            this.recordAverageValue = this.performanceConfiguration.reportAverage;
-            this.recordMedianValue = this.performanceConfiguration.reportMedian;
-            this.recordStdDevValue = this.performanceConfiguration.reportStandardDeviation;
-            this.recordMaxValue = this.performanceConfiguration.reportMax;
-            this.recordMinValue = this.performanceConfiguration.reportMin;
-            this.recordVariance = this.performanceConfiguration.reportVariance;
-        };
-        /**
-         * Gets a log stack trace.
-         */
-        TestCollector.prototype.getLogStackTrace = function () {
-            return new Error("Get stack trace.")
-                .stack
-                .toString()
-                .split("\n")
-                .slice(1)
-                .filter(wasmFilter)
-                .join("\n");
-        };
-        return TestCollector;
-    }());
-    exports.TestCollector = TestCollector;
-});
-define("test/TestContext", ["require", "exports", "reporter/DefaultTestReporter", "perf_hooks", "util/timeDifference", "test/TestCollector"], function (require, exports, DefaultTestReporter_1, perf_hooks_1, timeDifference_1, TestCollector_1) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    var TestContext = /** @class */ (function (_super) {
-        __extends(TestContext, _super);
-        function TestContext(props) {
-            var _this = _super.call(this, props) || this;
-            _this.time = 0;
-            _this.pass = true;
-            _this.startupTime = 0;
-            _this.reporter = new DefaultTestReporter_1.DefaultTestReporter();
-            /* istanbul ignore next */
-            _this.stdout = process ? process.stdout : null;
-            /* istanbul ignore next */
-            _this.stderr = process.stderr ? process.stderr : null;
-            _this.endGroup = false;
-            if (props) {
-                /* istanbul ignore next */
-                if (props.reporter)
-                    _this.reporter = props.reporter;
-                /* istanbul ignore next */
-                if (props.stdout)
-                    _this.stdout = props.stdout;
-                /* istanbul ignore next */
-                if (props.stderr)
-                    _this.stderr = props.stderr;
-            }
-            return _this;
-        }
-        /**
-         * Run the tests on the wasm module.
-         */
-        TestContext.prototype.run = function (wasm) {
-            // set wasm immediately
-            this.wasm = wasm;
-            // start the timer
-            var start = perf_hooks_1.performance.now();
-            // start the module up
-            _super.prototype.collectTests.call(this);
-            // calculate startuptime
-            this.startupTime = timeDifference_1.timeDifference(perf_hooks_1.performance.now(), start);
-            if (this.errors.length > 0)
-                return;
-            this.ready = true;
-            // start the test suite
-            this.reporter.onStart(this);
-            for (var _i = 0, _a = this.testGroups; _i < _a.length; _i++) {
-                var group = _a[_i];
-                this.runGroup(group);
-            }
-            var end = perf_hooks_1.performance.now();
-            this.time = timeDifference_1.timeDifference(end, start);
-            this.reporter.onFinish(this);
-        };
-        TestContext.prototype.runGroup = function (group) {
-            this.endGroup = false;
-            // set the group starttime
-            group.start = perf_hooks_1.performance.now();
-            // set the log target
-            this.logTarget = group;
-            // for each beforeAllCallback
-            this.runBeforeAll(group);
-            // report the group as started, and log all the beforeAll logs outside the describe block
-            this.reporter.onGroupStart(group);
-            if (this.endGroup)
-                return;
-            for (var _i = 0, _a = group.tests; _i < _a.length; _i++) {
-                var result = _a[_i];
-                this.runTest(group, result);
-                if (this.endGroup)
-                    return;
-                this.reporter.onTestFinish(group, result);
-                this.logTarget = group;
-            }
-            // for each afterAllCallback
-            this.runAfterAll(group);
-            if (this.endGroup)
-                return;
-            // finish the group
-            group.end = perf_hooks_1.performance.now();
-            group.time = timeDifference_1.timeDifference(group.end, group.start);
-            group.reason = "Test suite " + group.name + " passed successfully.";
-            this.reporter.onGroupFinish(group);
-        };
-        /**
-         * Run a given test.
-         *
-         * @param {RunContext} runContext - The current run context.
-         * @param {TestGroup} group - The current run group.
-         * @param {number} testIndex - The current test index.
-         */
-        TestContext.prototype.runTest = function (group, result) {
-            // set the log target
-            this.logTarget = result;
-            this.reporter.onTestStart(group, result);
-            result.ran = true;
-            result.start = perf_hooks_1.performance.now();
-            // If performance is enabled, use the performance values, otherwise, just run once.
-            if (result.performance) {
-                var runCount = 0;
-                var testStartTime = perf_hooks_1.performance.now();
-                var currentTestRunTime = 0;
-                // run the test loop
-                while (true) { // always run at least once
-                    this.runBeforeEach(group, result);
-                    /**
-                     * Especially because the performance functions are run repeatedly, if an error occurs, assume the
-                     * worst and skip the test group. These functions definitely are assumed to be safe by the test context.
-                     */
-                    if (this.endGroup)
-                        return;
-                    this.runTestCall(group, result);
-                    this.runAfterEach(group, result);
-                    if (this.endGroup)
-                        return; // check to see if the afterEach functions errored (see above)
-                    currentTestRunTime = perf_hooks_1.performance.now() - testStartTime; // calculate how long the current test has run
-                    runCount += 1; // increase the run count
-                    if (runCount >= result.maxSamples)
-                        break; // if we have reached the max sample count
-                    if (currentTestRunTime >= result.maxRuntime)
-                        break; // weve collected enough samples and the test is over
-                }
-                if (result.calculateAverageValue)
-                    result.calculateAverage();
-                if (result.calculateMaxValue)
-                    result.calculateMax();
-                if (result.calculateMedianValue)
-                    result.calculateMedian();
-                if (result.calculateMinValue)
-                    result.calculateMin();
-                if (result.calculateVarianceValue)
-                    result.calculateVariance();
-                if (result.calculateStandardDeviationValue)
-                    result.calculateStandardDeviation();
-            }
-            else {
-                this.runBeforeEach(group, result);
-                if (this.endGroup)
-                    return;
-                this.runTestCall(group, result);
-                this.runAfterEach(group, result);
-                if (this.endGroup)
-                    return;
-            }
-            result.end = perf_hooks_1.performance.now();
-            result.runTime = timeDifference_1.timeDifference(result.end, result.start);
-        };
-        /**
-         * Run the current test once and collect statistics.
-         *
-         * @param {RunContext} runContext - The current run context.
-         * @param {TestGroup} group - The current test group.
-         * @param {TestResult} result - The current test result.
-         * @param {number} testIndex - The current test index.
-         */
-        TestContext.prototype.runTestCall = function (group, result) {
-            var start = perf_hooks_1.performance.now();
-            var testCallResult = this.tryCall(result.functionPointer);
-            var end = perf_hooks_1.performance.now();
-            result.times.push(timeDifference_1.timeDifference(end, start));
-            result.pass = result.negated
-                ? (testCallResult === 0)
-                : (testCallResult === 1);
-            if (!result.pass) {
-                group.pass = false;
-                // if it's not negated then set the message, the actual, expected, and stack values
-                if (!result.negated) {
-                    result.message = this.message;
-                    result.actual = this.actual;
-                    result.expected = this.expected;
-                    result.stack = this.stack;
-                }
-            }
-        };
-        /**
-         * Run the afterEach callbacks before running the test.
-         *
-         * @param {RunContext} runContext - The current run context.
-         * @param {TestGroup} group - The current test group.
-         * @param {TestResult} result - The current test result.
-         */
-        TestContext.prototype.runAfterEach = function (group, result) {
-            // for each afterEach callback function pointer
-            for (var _i = 0, _a = group.afterEachPointers; _i < _a.length; _i++) {
-                var afterEachCallback = _a[_i];
-                var afterEachResult = this.tryCall(afterEachCallback);
-                // if afterEach fails
-                if (afterEachResult === 0) {
-                    this.endGroup = true;
-                    group.end = result.end = perf_hooks_1.performance.now();
-                    group.pass = false;
-                    this.pass = false;
-                    group.reason = "Test suite " + group.name + " failed in afterEach callback.";
-                    result.pass = false;
-                    group.time = timeDifference_1.timeDifference(group.end, group.start);
-                    this.reporter.onTestFinish(group, result);
-                    this.reporter.onGroupFinish(group);
-                    return;
-                }
-            }
-        };
-        /**
-         * Run the beforeEach callbacks before running the test.
-         *
-         * @param {RunContext} runContext - The current run context.
-         * @param {TestGroup} group - The current test group.
-         * @param {TestResult} result - The current test result.
-         */
-        TestContext.prototype.runBeforeEach = function (group, result) {
-            // for each beforeEach callback function pointer
-            for (var _i = 0, _a = group.beforeEachPointers; _i < _a.length; _i++) {
-                var beforeEachCallback = _a[_i];
-                var beforeEachResult = this.tryCall(beforeEachCallback);
-                // if beforeEach fails
-                if (beforeEachResult === 0) {
-                    result.end = group.end = perf_hooks_1.performance.now();
-                    group.pass = false;
-                    group.reason = group.reason = "Test suite " + group.name + " failed in beforeEach callback.";
-                    result.pass = false;
-                    group.time = timeDifference_1.timeDifference(group.end, group.start);
-                    this.reporter.onTestFinish(group, result);
-                    this.reporter.onGroupFinish(group);
-                    this.endGroup = true;
-                    return;
-                }
-            }
-        };
-        /**
-         * Run the afterAll callbacks with the given runContext and group.
-         *
-         * @param {RunContext} runContext - The current run context.
-         * @param {TestGroup} group - The current test group.
-         */
-        TestContext.prototype.runAfterAll = function (group) {
-            for (var _i = 0, _a = group.afterAllPointers; _i < _a.length; _i++) {
-                var afterAllCallback = _a[_i];
-                // call each afterAll callback
-                var afterAllResult = this.tryCall(afterAllCallback);
-                // if the test fails
-                if (afterAllResult === 0) {
-                    group.end = perf_hooks_1.performance.now();
-                    group.pass = false;
-                    group.reason = "Test suite " + group.name + " failed in afterAll callback.";
-                    this.pass = false;
-                    group.time = timeDifference_1.timeDifference(group.end, group.start);
-                    this.reporter.onGroupFinish(group);
-                    this.endGroup = true;
-                    return;
-                }
-            }
-        };
-        /**
-         * Run the beforeAll callbacks with the given runContext and group.
-         *
-         * @param {RunContext} runContext - The current run context.
-         * @param {TestGroup} group - The current test group.
-         */
-        TestContext.prototype.runBeforeAll = function (group) {
-            for (var _i = 0, _a = group.beforeAllPointers; _i < _a.length; _i++) {
-                var beforeAllCallback = _a[_i];
-                // call each beforeAll callback
-                var beforeAllResult = this.tryCall(beforeAllCallback);
-                // if the test fails
-                if (beforeAllResult === 0) {
-                    group.end = perf_hooks_1.performance.now();
-                    group.pass = false;
-                    group.reason = "Test suite " + group.name + " failed in beforeAll callback.";
-                    this.pass = false;
-                    group.time = timeDifference_1.timeDifference(group.end, group.start);
-                    this.endGroup = true;
-                    return;
-                }
-            }
-        };
-        return TestContext;
-    }(TestCollector_1.TestCollector));
-    exports.TestContext = TestContext;
-});
-define("reporter/CSVTestReporter", ["require", "exports", "test/TestReporter", "csv-stringify", "fs", "path"], function (require, exports, TestReporter_2, csv_stringify_1, fs_1, path_1) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    csv_stringify_1 = __importDefault(csv_stringify_1);
-    var columns = ["Group", "Name", "Ran", "Pass", "Runtime", "Message", "Actual", "Expected", "Average", "Mean", "Max", "Min", "StdDev", "Variance"];
-    var CSVTestReporter = /** @class */ (function (_super) {
-        __extends(CSVTestReporter, _super);
-        function CSVTestReporter() {
-            var _this = _super !== null && _super.apply(this, arguments) || this;
-            _this.output = null;
-            _this.fileName = null;
-            return _this;
-        }
-        CSVTestReporter.prototype.onStart = function (suite) {
-            this.output = csv_stringify_1.default({ columns: columns });
-            var extension = path_1.extname(suite.fileName);
-            var dir = path_1.dirname(suite.fileName);
-            var base = path_1.basename(suite.fileName, extension);
-            var outPath = path_1.join(process.cwd(), dir, base + ".csv");
-            this.fileName = fs_1.createWriteStream(outPath, "utf8");
-            this.output.pipe(this.fileName);
-            this.output.write(columns);
-        };
-        CSVTestReporter.prototype.onGroupStart = function () { };
-        CSVTestReporter.prototype.onGroupFinish = function () { };
-        CSVTestReporter.prototype.onFinish = function () {
-            this.output.end();
-        };
-        CSVTestReporter.prototype.onTestStart = function () { };
-        CSVTestReporter.prototype.onTestFinish = function (group, result) {
-            this.output.write([
-                group.name,
-                result.ran ? "RAN" : "NOT RUN",
-                result.name,
-                result.pass ? "PASS" : "FAIL",
-                result.runTime.toString(),
-                result.message,
-                result.actual ? result.actual.message : "",
-                result.expected ? result.expected.message : "",
-                result.hasAverage ? result.average.toString() : "",
-                result.hasMedian ? result.median.toString() : "",
-                result.hasMax ? result.max.toString() : "",
-                result.hasMin ? result.min.toString() : "",
-                result.hasStdDev ? result.stdDev.toString() : "",
-                result.hasVariance ? result.variance.toString() : ""
-            ]);
-        };
-        CSVTestReporter.prototype.onTodo = function (group, desc) {
-            this.output.write([
-                group.name,
-                "TODO",
-                desc,
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-            ]);
-        };
-        return CSVTestReporter;
-    }(TestReporter_2.TestReporter));
-    exports.CSVTestReporter = CSVTestReporter;
-});
-define("reporter/EmptyReporter", ["require", "exports", "test/TestReporter"], function (require, exports, TestReporter_3) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    var EmptyReporter = /** @class */ (function (_super) {
-        __extends(EmptyReporter, _super);
-        function EmptyReporter() {
-            return _super !== null && _super.apply(this, arguments) || this;
-        }
-        EmptyReporter.prototype.onFinish = function () { };
-        EmptyReporter.prototype.onGroupFinish = function () { };
-        EmptyReporter.prototype.onGroupStart = function () { };
-        EmptyReporter.prototype.onStart = function () { };
-        EmptyReporter.prototype.onTestFinish = function () { };
-        EmptyReporter.prototype.onTestStart = function () { };
-        EmptyReporter.prototype.onTodo = function () { };
-        return EmptyReporter;
-    }(TestReporter_3.TestReporter));
-    exports.EmptyReporter = EmptyReporter;
 });
 define("reporter/JSONTestReporter", ["require", "exports", "test/TestReporter", "fs", "path"], function (require, exports, TestReporter_4, fs_2, path_2) {
     "use strict";
@@ -1838,6 +1202,11 @@ define("cli/help", ["require", "exports", "chalk"], function (require, exports, 
     exports.help = help;
     var templateObject_42;
 });
+define("util/timeDifference", ["require", "exports"], function (require, exports) {
+    "use strict";
+    Object.defineProperty(exports, "__esModule", { value: true });
+    exports.timeDifference = function (end, start) { return Math.round((end - start) * 1000) / 1000; };
+});
 define("cli/util/IYargs", ["require", "exports"], function (require, exports) {
     "use strict";
     Object.defineProperty(exports, "__esModule", { value: true });
@@ -1872,7 +1241,7 @@ define("cli/util/collectPerformanceConfiguration", ["require", "exports"], funct
     }
     exports.collectPerformanceConfiguration = collectPerformanceConfiguration;
 });
-define("cli/util/collectReporter", ["require", "exports", "reporter/DefaultTestReporter", "reporter/EmptyReporter", "reporter/SummaryTestReporter", "path", "reporter/CSVTestReporter", "reporter/JSONTestReporter"], function (require, exports, DefaultTestReporter_2, EmptyReporter_1, SummaryTestReporter_1, path_4, CSVTestReporter_1, JSONTestReporter_1) {
+define("cli/util/collectReporter", ["require", "exports", "reporter/DefaultTestReporter", "reporter/EmptyReporter", "reporter/SummaryTestReporter", "path", "reporter/CSVTestReporter", "reporter/JSONTestReporter"], function (require, exports, DefaultTestReporter_1, EmptyReporter_2, SummaryTestReporter_1, path_4, CSVTestReporter_1, JSONTestReporter_1) {
     "use strict";
     Object.defineProperty(exports, "__esModule", { value: true });
     path_4 = __importDefault(path_4);
@@ -1895,7 +1264,7 @@ define("cli/util/collectReporter", ["require", "exports", "reporter/DefaultTestR
                     }
                 }
                 else {
-                    return new DefaultTestReporter_2.DefaultTestReporter();
+                    return new DefaultTestReporter_1.DefaultTestReporter();
                 }
             }
             catch (ex) {
@@ -1907,7 +1276,7 @@ define("cli/util/collectReporter", ["require", "exports", "reporter/DefaultTestR
             }
         }
         else if (targetReporter === "EmptyReporter") {
-            return new EmptyReporter_1.EmptyReporter();
+            return new EmptyReporter_2.EmptyReporter();
         }
         else if (targetReporter === "SummaryTestReporter") {
             return new SummaryTestReporter_1.SummaryTestReporter();
@@ -1919,7 +1288,7 @@ define("cli/util/collectReporter", ["require", "exports", "reporter/DefaultTestR
             return new JSONTestReporter_1.JSONTestReporter();
         }
         else {
-            return new DefaultTestReporter_2.DefaultTestReporter();
+            return new DefaultTestReporter_1.DefaultTestReporter();
         }
     }
     exports.collectReporter = collectReporter;
@@ -1969,7 +1338,7 @@ define("cli/util/writeFile", ["require", "exports", "fs"], function (require, ex
     }
     exports.writeFile = writeFile;
 });
-define("cli/run", ["require", "exports", "assemblyscript/dist/asc", "assemblyscript/cli/util/options", "test/TestContext", "fs", "assemblyscript/lib/loader", "reporter/DefaultTestReporter", "perf_hooks", "util/timeDifference", "util/IPerformanceConfiguration", "path", "chalk", "glob", "cli/util/collectPerformanceConfiguration", "cli/util/collectReporter", "cli/util/getTestEntryFiles", "cli/util/writeFile"], function (require, exports, asc_1, options_1, TestContext_1, fs, loader_1, DefaultTestReporter_3, perf_hooks_2, timeDifference_2, IPerformanceConfiguration_2, path, chalk_6, glob_2, collectPerformanceConfiguration_1, collectReporter_1, getTestEntryFiles_1, writeFile_1) {
+define("cli/run", ["require", "exports", "assemblyscript/dist/asc", "assemblyscript/cli/util/options", "test/TestContext", "fs", "assemblyscript/lib/loader", "reporter/DefaultTestReporter", "perf_hooks", "util/timeDifference", "util/IPerformanceConfiguration", "path", "chalk", "glob", "cli/util/collectPerformanceConfiguration", "cli/util/collectReporter", "cli/util/getTestEntryFiles", "cli/util/writeFile"], function (require, exports, asc_1, options_1, TestContext_1, fs, loader_1, DefaultTestReporter_2, perf_hooks_1, timeDifference_1, IPerformanceConfiguration_1, path, chalk_6, glob_2, collectPerformanceConfiguration_1, collectReporter_1, getTestEntryFiles_1, writeFile_1) {
     "use strict";
     Object.defineProperty(exports, "__esModule", { value: true });
     asc_1 = __importDefault(asc_1);
@@ -1978,7 +1347,7 @@ define("cli/run", ["require", "exports", "assemblyscript/dist/asc", "assemblyscr
     chalk_6 = __importDefault(chalk_6);
     glob_2 = __importDefault(glob_2);
     function run(yargs, compilerArgs) {
-        var start = perf_hooks_2.performance.now();
+        var start = perf_hooks_1.performance.now();
         // obtain the configuration file
         var configurationPath = path.resolve(process.cwd(), yargs.argv.c || yargs.argv.config || "./as-pect.config.js");
         console.log(chalk_6.default(templateObject_43 || (templateObject_43 = __makeTemplateObject(["{bgWhite.black [Log]} Using configuration ", ""], ["{bgWhite.black [Log]} Using configuration ", ""])), configurationPath));
@@ -2015,8 +1384,8 @@ define("cli/run", ["require", "exports", "assemblyscript/dist/asc", "assemblyscr
         // if a reporter is specified in cli arguments, override configuration
         var reporter = (yargs.argv.reporter || yargs.argv.r)
             ? collectReporter_1.collectReporter(yargs)
-            : configuration.reporter || new DefaultTestReporter_3.DefaultTestReporter();
-        var performanceConfiguration = configuration.performance || IPerformanceConfiguration_2.createDefaultPerformanceConfiguration();
+            : configuration.reporter || new DefaultTestReporter_2.DefaultTestReporter();
+        var performanceConfiguration = configuration.performance || IPerformanceConfiguration_1.createDefaultPerformanceConfiguration();
         // setup performance options, overriding configured values if the flag is passed to the cli
         collectPerformanceConfiguration_1.collectPerformanceConfiguration(yargs, performanceConfiguration);
         // include all the file globs
@@ -2174,7 +1543,7 @@ define("cli/run", ["require", "exports", "assemblyscript/dist/asc", "assemblyscr
                 // if any tests failed, and they all ran, exit(1)
                 if (count === 0) {
                     if (runTests) {
-                        var end = perf_hooks_2.performance.now();
+                        var end = perf_hooks_1.performance.now();
                         failed = testCount !== successCount || errors.length > 0;
                         var result = failed
                             ? chalk_6.default(templateObject_53 || (templateObject_53 = __makeTemplateObject(["{red \u2716 FAIL}"], ["{red \u2716 FAIL}"]))) : chalk_6.default(templateObject_54 || (templateObject_54 = __makeTemplateObject(["{green \u2714 PASS}"], ["{green \u2714 PASS}"])));
@@ -2183,7 +1552,7 @@ define("cli/run", ["require", "exports", "assemblyscript/dist/asc", "assemblyscr
                             var error_1 = errors_1[_i];
                             console.log(chalk_6.default(templateObject_55 || (templateObject_55 = __makeTemplateObject(["\n   [Error]: {red ", "}: ", "\n   [Stack]: {yellow ", "}\n  "], ["\n   [Error]: {red ", "}: ", "\n   [Stack]: {yellow ", "}\n  "])), error_1.type, error_1.message, error_1.stackTrace.split("\n").join("\n            ")));
                         }
-                        console.log("\n  [Result]: " + result + "\n   [Files]: " + testEntryFiles.size + " total\n  [Groups]: " + groupCount + " count, " + groupSuccessCount + " pass\n   [Tests]: " + successCount.toString() + " pass, " + (testCount - successCount).toString() + " fail, " + testCount.toString() + " total\n    [Time]: " + timeDifference_2.timeDifference(end, start).toString() + "ms");
+                        console.log("\n  [Result]: " + result + "\n   [Files]: " + testEntryFiles.size + " total\n  [Groups]: " + groupCount + " count, " + groupSuccessCount + " pass\n   [Tests]: " + successCount.toString() + " pass, " + (testCount - successCount).toString() + " fail, " + testCount.toString() + " total\n    [Time]: " + timeDifference_1.timeDifference(end, start).toString() + "ms");
                     }
                     Promise.all(filePromises).then(function () {
                         if (failed)
@@ -2250,23 +1619,23 @@ define("cli/index", ["require", "exports", "chalk", "path", "yargs-parser", "cli
     exports.asp = asp;
     var templateObject_56;
 });
-define("as-pect", ["require", "exports", "test/TestContext", "test/TestGroup", "test/TestReporter", "test/TestResult", "reporter/CSVTestReporter", "reporter/DefaultTestReporter", "reporter/EmptyReporter", "reporter/JSONTestReporter", "reporter/SummaryTestReporter", "util/ActualValue", "util/IPerformanceConfiguration", "util/LogValue", "cli/index"], function (require, exports, TestContext_2, TestGroup_2, TestReporter_6, TestResult_2, CSVTestReporter_2, DefaultTestReporter_4, EmptyReporter_2, JSONTestReporter_2, SummaryTestReporter_2, ActualValue_2, IPerformanceConfiguration_3, LogValue_3, cli_1) {
+define("as-pect", ["require", "exports", "test/TestContext", "test/TestGroup", "test/TestReporter", "test/TestResult", "reporter/CSVTestReporter", "reporter/DefaultTestReporter", "reporter/EmptyReporter", "reporter/JSONTestReporter", "reporter/SummaryTestReporter", "util/ActualValue", "util/IPerformanceConfiguration", "util/LogValue", "cli/index"], function (require, exports, TestContext_2, TestGroup_1, TestReporter_6, TestResult_1, CSVTestReporter_2, DefaultTestReporter_3, EmptyReporter_3, JSONTestReporter_2, SummaryTestReporter_2, ActualValue_2, IPerformanceConfiguration_2, LogValue_3, cli_1) {
     "use strict";
     function __export(m) {
         for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
     }
     Object.defineProperty(exports, "__esModule", { value: true });
     __export(TestContext_2);
-    __export(TestGroup_2);
+    __export(TestGroup_1);
     __export(TestReporter_6);
-    __export(TestResult_2);
+    __export(TestResult_1);
     __export(CSVTestReporter_2);
-    __export(DefaultTestReporter_4);
-    __export(EmptyReporter_2);
+    __export(DefaultTestReporter_3);
+    __export(EmptyReporter_3);
     __export(JSONTestReporter_2);
     __export(SummaryTestReporter_2);
     __export(ActualValue_2);
-    __export(IPerformanceConfiguration_3);
+    __export(IPerformanceConfiguration_2);
     __export(LogValue_3);
     __export(cli_1);
 });
