--- lib/test/TestCollector.js
@@ -1,747 +1,0 @@
-(function (factory) {
-    if (typeof module === "object" && typeof module.exports === "object") {
-        var v = factory(require, exports);
-        if (v !== undefined) module.exports = v;
-    }
-    else if (typeof define === "function" && define.amd) {
-        define(["require", "exports", "../util/LogValue", "../util/ActualValue", "./TestGroup", "../util/IPerformanceConfiguration", "./TestResult"], factory);
-    }
-})(function (require, exports) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    var LogValue_1 = require("../util/LogValue");
-    var ActualValue_1 = require("../util/ActualValue");
-    var TestGroup_1 = require("./TestGroup");
-    var IPerformanceConfiguration_1 = require("../util/IPerformanceConfiguration");
-    var TestResult_1 = require("./TestResult");
-    var wasmFilter = function (input) { return /wasm-function/i.test(input); };
-    /**
-     * This class is responsible for collecting all the tests in a test binary.
-     */
-    var TestCollector = /** @class */ (function () {
-        function TestCollector(props) {
-            this.wasm = null;
-            // test group values
-            this.groupStack = [new TestGroup_1.TestGroup()];
-            this.testGroups = [];
-            this.logTarget = this.groupStack[0];
-            // public warning/error lists
-            this.errors = [];
-            // public fileName
-            this.fileName = "";
-            // test state machine values
-            this.stack = "";
-            this.message = "";
-            this.actual = null;
-            this.expected = null;
-            this.performanceConfiguration = IPerformanceConfiguration_1.createDefaultPerformanceConfiguration();
-            /**
-             * This value is used to detect if an `expect()` function call was used outside of a test
-             * function. If a reportExpected or reportActual function is called before the `context.run()`
-             * method is called, it should prevent the `run()` method from running the tests and report a
-             * failure.
-             */
-            this.ready = false;
-            /**
-             * These are the test and group filters for the binary. They must be provided before collection
-             * begins.
-             */
-            this.testRegex = new RegExp("");
-            this.groupRegex = new RegExp("");
-            if (props) {
-                /* istanbul ignore next */
-                if (props.fileName)
-                    this.fileName = props.fileName;
-                /* istanbul ignore next */
-                if (props.testRegex)
-                    this.testRegex = props.testRegex;
-                /* istanbul ignore next */
-                if (props.groupRegex)
-                    this.groupRegex = props.groupRegex;
-                /* istanbul ignore next */
-                if (props.performanceConfiguration)
-                    this.performanceConfiguration = props.performanceConfiguration;
-            }
-        }
-        /**
-         * Call this method to start the `__main()` method provided by the `as-pect` exports to start the
-         * process of test collection.
-         */
-        TestCollector.prototype.collectTests = function () {
-            // reset the performance values first, then collect the tests by calling `__main()`
-            this.resetPerformanceValues();
-            this.wasm.__main();
-        };
-        /**
-         * This method creates a WebAssembly imports object with all the TestContext functions
-         * bound to the TestContext.
-         *
-         * @param {any[]} imports - Every import item specified.
-         */
-        TestCollector.prototype.createImports = function () {
-            var _this = this;
-            var imports = [];
-            for (var _i = 0; _i < arguments.length; _i++) {
-                imports[_i] = arguments[_i];
-            }
-            var result = Object.assign.apply(Object, [{}].concat(imports, [{
-                    __aspect: {
-                        clearExpected: this.clearExpected.bind(this),
-                        debug: this.debug.bind(this),
-                        tryCall: this.tryCall.bind(this),
-                        logNull: this.logNull.bind(this),
-                        logReference: this.logReference.bind(this),
-                        logString: this.logString.bind(this),
-                        logValue: this.logValue.bind(this),
-                        reportDescribe: this.reportDescribe.bind(this),
-                        reportEndDescribe: this.reportEndDescribe.bind(this),
-                        reportTest: this.reportTest.bind(this),
-                        reportBeforeEach: this.reportBeforeEach.bind(this),
-                        reportBeforeAll: this.reportBeforeAll.bind(this),
-                        reportAfterEach: this.reportAfterEach.bind(this),
-                        reportAfterAll: this.reportAfterAll.bind(this),
-                        reportTodo: this.reportTodo.bind(this),
-                        reportActualNull: this.reportActualNull.bind(this),
-                        reportExpectedNull: this.reportExpectedNull.bind(this),
-                        reportActualValue: this.reportActualValue.bind(this),
-                        reportExpectedValue: this.reportExpectedValue.bind(this),
-                        reportActualReference: this.reportActualReference.bind(this),
-                        reportExpectedReference: this.reportExpectedReference.bind(this),
-                        reportActualString: this.reportActualString.bind(this),
-                        reportExpectedString: this.reportExpectedString.bind(this),
-                        reportExpectedTruthy: this.reportExpectedTruthy.bind(this),
-                        reportExpectedFalsy: this.reportExpectedFalsy.bind(this),
-                        reportExpectedFinite: this.reportExpectedFinite.bind(this),
-                        reportNegatedTest: this.reportNegatedTest.bind(this),
-                        performanceEnabled: this.performanceEnabled.bind(this),
-                        maxSamples: this.maxSamples.bind(this),
-                        maxTestRunTime: this.maxTestRunTime.bind(this),
-                        roundDecimalPlaces: this.roundDecimalPlaces.bind(this),
-                        reportAverage: this.reportAverage.bind(this),
-                        reportMedian: this.reportMedian.bind(this),
-                        reportStdDev: this.reportStdDev.bind(this),
-                        reportMax: this.reportMax.bind(this),
-                        reportMin: this.reportMin.bind(this),
-                        reportVariance: this.reportVariance.bind(this),
-                    },
-                }]));
-            result.env = result.env || {};
-            var previousAbort = (result.env.abort) || (function () { });
-            result.env.abort = function () {
-                var args = [];
-                for (var _i = 0; _i < arguments.length; _i++) {
-                    args[_i] = arguments[_i];
-                }
-                previousAbort.apply(void 0, args);
-                // @ts-ignore
-                _this.abort.apply(_this, args);
-            };
-            return result;
-        };
-        /**
-         * This is called to stop the debugger.  e.g. `node --inspect-brk asp`.
-         */
-        TestCollector.prototype.debug = function () { debugger; };
-        /**
-         * This is a web assembly utility function that wraps a function call in a try catch block to
-         * report success or failure.
-         *
-         * @param {number} pointer - The function pointer to call. It must accept no parameters and return
-         * void.
-         * @returns {1 | 0} - If the callback was run successfully without error, it returns 1, else it
-         * returns 0.
-         */
-        TestCollector.prototype.tryCall = function (pointer) {
-            if (pointer === -1)
-                return 1;
-            try {
-                this.wasm.__call(pointer);
-            }
-            catch (ex) {
-                this.stack = this.getErrorStackTrace(ex);
-                return 0;
-            }
-            return 1;
-        };
-        /**
-         * Log a null value to the reporter.
-         */
-        TestCollector.prototype.logNull = function () {
-            // create a new log value
-            var value = new LogValue_1.LogValue();
-            var target = this.logTarget;
-            // collect log metadata
-            value.stack = this.getLogStackTrace();
-            value.message = "null";
-            value.target = target;
-            // push the log value to the logs
-            target.logs.push(value);
-        };
-        /**
-         * This function is called after each expectation if the expectation passes. This prevents other
-         * unreachable() conditions that throw errors to report actual and expected values too.
-         */
-        TestCollector.prototype.clearExpected = function () {
-            this.expected = null;
-            this.actual = null;
-            this.stack = "";
-        };
-        /**
-         * Log a reference to the reporter.
-         *
-         * @param {number} referencePointer - The pointer to the reference.
-         * @param {number} offset - The offset of the reference.
-         */
-        TestCollector.prototype.logReference = function (referencePointer, offset) {
-            var value = new LogValue_1.LogValue();
-            var target = this.logTarget;
-            value.bytes = Array.from(this.wasm.U8.slice(referencePointer, referencePointer + offset));
-            value.message = "Reference Type";
-            value.offset = offset;
-            value.pointer = referencePointer;
-            value.stack = this.getLogStackTrace();
-            value.target = target;
-            value.value = referencePointer;
-            // push the log value to the logs
-            target.logs.push(value);
-        };
-        /**
-         * This adds a logged string to the current test.
-         *
-         * @param {number} pointer - The pointer to the logged string reference.
-         */
-        TestCollector.prototype.logString = function (pointer) {
-            var value = new LogValue_1.LogValue();
-            var target = this.logTarget;
-            value.message = this.wasm.__getString(pointer);
-            value.offset = 0;
-            value.pointer = pointer;
-            value.stack = this.getLogStackTrace();
-            value.target = target;
-            value.value = pointer;
-            // push the log value to the logs
-            target.logs.push(value);
-        };
-        /**
-         * Log a numevalueric value to the reporter.
-         *
-         * @param {number} value - The value to be logged.
-         */
-        TestCollector.prototype.logValue = function (numericValue) {
-            var value = new LogValue_1.LogValue();
-            var target = this.logTarget;
-            value.stack = this.getLogStackTrace();
-            value.message = "Value " + numericValue.toString();
-            value.value = numericValue;
-            value.target = target;
-            // push the log value to the logs
-            target.logs.push(value);
-        };
-        /**
-         * This web assembly linked function creates a test group. It's called when the test suite calls
-         * the describe("test", callback) function from within AssemblyScript. It receives a pointer to
-         * the description of the tests, forks the top level test group, pushes the suiteName to a list,
-         * then pushes the forked group to the top of the test context stack.
-         *
-         * @param {number} suiteNamePointer
-         */
-        TestCollector.prototype.reportDescribe = function (suiteNamePointer) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            var nextGroup = group.fork();
-            nextGroup.name = group.name + this.wasm.__getString(suiteNamePointer);
-            nextGroup.willRun = this.groupRegex.test(nextGroup.name);
-            this.groupStack.push(nextGroup);
-            this.logTarget = nextGroup;
-        };
-        /**
-         * This web assembly linked function finishes a test group. It's called when the test suite calls
-         * the describe("test", callback) function from within AssemblyScript. It pops the current
-         * test group from the test context stack and pushes it to the final test group list.
-         */
-        TestCollector.prototype.reportEndDescribe = function () {
-            var next = this.groupStack.pop();
-            var testCount = next.tests.length;
-            /**
-             * If a describe finishes first, it happens BEFORE other describes. This means
-             * inner describe blocks run at lower priority than outer describe blocks. It also should be
-             * pushed if it willRun.
-             */
-            if (next.willRun && testCount > 0) {
-                this.testGroups.unshift(next);
-            }
-            this.logTarget = this.groupStack[this.groupStack.length - 1];
-        };
-        /**
-         * This web assembly linked function sets the group's "beforeEach" callback pointer to
-         * the current groupStackItem.
-         *
-         * @param {number} callbackPointer - The callback that should run before each test.
-         */
-        TestCollector.prototype.reportBeforeEach = function (callbackPointer) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            ;
-            group.beforeEachPointers.push(callbackPointer);
-        };
-        /**
-         * This web assembly linked function adds the group's "beforeAll" callback pointer to
-         * the current groupStackItem.
-         *
-         * @param {number} callbackPointer - The callback that should run before each test in the
-         * current context.
-         */
-        TestCollector.prototype.reportBeforeAll = function (callbackPointer) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            ;
-            group.beforeAllPointers.push(callbackPointer);
-        };
-        /**
-         * This web assembly linked function sets the group's "afterEach" callback pointer.
-         *
-         * @param {number} callbackPointer - The callback that should run before each test group.
-         */
-        TestCollector.prototype.reportAfterEach = function (callbackPointer) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            ;
-            group.afterEachPointers.push(callbackPointer);
-        };
-        /**
-         * This web assembly linked function adds the group's "afterAll" callback pointer to
-         * the current groupStackItem.
-         *
-         * @param {number} callbackPointer - The callback that should run before each test in the
-         * current context.
-         */
-        TestCollector.prototype.reportAfterAll = function (callbackPointer) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            ;
-            group.afterAllPointers.push(callbackPointer);
-        };
-        /**
-         * This web assembly linked function creates a test from the callback and the testNamePointer in
-         * the current group. It assumes that the group has already been created with the describe
-         * function. It is called when `it("description", callback)` or `test("description", callback)`
-         * is called.
-         *
-         * @param {number} testNamePointer - The test's name pointer.
-         * @param {number} callback - The test's function.
-         */
-        TestCollector.prototype.reportTest = function (testNamePointer, callback) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            if (!group.willRun)
-                return;
-            var name = this.wasm.__getString(testNamePointer);
-            if (!this.testRegex.test(name))
-                return;
-            var test = new TestResult_1.TestResult();
-            test.functionPointer = callback;
-            test.name = name;
-            test.message = "";
-            test.negated = false;
-            test.performance = this.performanceEnabledValue || false;
-            if (test.performance) {
-                test.maxSamples = !isFinite(this.maxSamplesValue)
-                    ? 10000 /* MaxSamples */
-                    : Math.min(this.maxSamplesValue, 10000 /* MaxSamples */);
-                test.maxRuntime = !isFinite(this.maxTestRunTimeValue)
-                    ? 5000 /* MaxTestRuntime */
-                    : Math.min(this.maxTestRunTimeValue, 5000 /* MaxTestRuntime */);
-                test.decimalPlaces = !isFinite(this.roundDecimalPlacesValue)
-                    ? 3
-                    : Math.max(Math.round(this.roundDecimalPlacesValue), 0 /* MinimumDecimalPlaces */);
-                test.calculateAverageValue = this.recordAverageValue || false;
-                test.calculateMedianValue = this.recordMedianValue || false;
-                test.calculateStandardDeviationValue = this.recordStdDevValue || false;
-                test.calculateMaxValue = this.recordMaxValue || false;
-                test.calculateMinValue = this.recordMinValue || false;
-                test.calculateVarianceValue = this.recordVariance || false;
-            }
-            group.tests.push(test);
-            this.resetPerformanceValues();
-        };
-        /**
-         * This web assembly linked function is responsible for reporting tests that are expected
-         * to fail. This is useful for verifying that specific application states will throw.
-         *
-         * @param {number} testNamePointer - The test's name pointer.
-         * @param {number} callback - The test's function.
-         * @param {number} message - The message associated with this test if it does not throw.
-         */
-        TestCollector.prototype.reportNegatedTest = function (testNamePointer, callback, message) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            if (!group.willRun)
-                return;
-            var name = this.wasm.__getString(testNamePointer);
-            if (!this.testRegex.test(name))
-                return;
-            var test = new TestResult_1.TestResult();
-            test.functionPointer = callback;
-            test.name = name;
-            test.message = this.wasm.__getString(message);
-            test.negated = true;
-            test.performance = this.performanceEnabledValue || false;
-            /* istanbul ignore next */
-            if (test.performance) {
-                /* istanbul ignore next */
-                test.maxSamples = !isFinite(this.maxSamplesValue)
-                    ? 10000 /* MaxSamples */
-                    : Math.min(this.maxSamplesValue, 10000 /* MaxSamples */);
-                /* istanbul ignore next */
-                test.maxRuntime = !isFinite(this.maxTestRunTimeValue)
-                    ? 5000 /* MaxTestRuntime */
-                    : Math.min(this.maxTestRunTimeValue, 5000 /* MaxTestRuntime */);
-                /* istanbul ignore next */
-                test.decimalPlaces = !isFinite(this.roundDecimalPlacesValue)
-                    ? 3
-                    : Math.max(Math.round(this.roundDecimalPlacesValue), 0 /* MinimumDecimalPlaces */);
-                /* istanbul ignore next */
-                test.calculateAverageValue = this.recordAverageValue || false;
-                /* istanbul ignore next */
-                test.calculateMedianValue = this.recordMedianValue || false;
-                /* istanbul ignore next */
-                test.calculateStandardDeviationValue = this.recordStdDevValue || false;
-                /* istanbul ignore next */
-                test.calculateMaxValue = this.recordMaxValue || false;
-                /* istanbul ignore next */
-                test.calculateMinValue = this.recordMaxValue || false;
-                /* istanbul ignore next */
-                test.calculateVarianceValue = this.recordVariance || false;
-            }
-            group.tests.push(test);
-            this.resetPerformanceValues();
-        };
-        /**
-         * This function reports a single "todo" item in a test suite.
-         *
-         * @param {number} todoPointer - The todo description string pointer.
-         */
-        TestCollector.prototype.reportTodo = function (todoPointer) {
-            var group = this.groupStack[this.groupStack.length - 1];
-            group.todos.push(this.wasm.__getString(todoPointer));
-        };
-        /**
-         * This function reports an actual null value.
-         */
-        TestCollector.prototype.reportActualNull = function () {
-            if (!this.ready) {
-                this.reportInvalidExpectCall();
-                return;
-            }
-            var value = new ActualValue_1.ActualValue();
-            value.message = "null";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.value = null;
-            this.actual = value;
-        };
-        /**
-         * This function reports an expected null value.
-         *
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedNull = function (negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = "null";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.negated = negated === 1;
-            value.value = null;
-            this.expected = value;
-        };
-        /**
-         * This function reports an actual numeric value.
-         *
-         * @param {number} numericValue - The value to be expected.
-         */
-        TestCollector.prototype.reportActualValue = function (numericValue) {
-            if (!this.ready) {
-                this.reportInvalidExpectCall();
-                return;
-            }
-            var value = new ActualValue_1.ActualValue();
-            value.message = numericValue.toString();
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.value = numericValue;
-            this.actual = value;
-        };
-        /**
-         * This function reports an expected numeric value.
-         *
-         * @param {number} numericValue - The value to be expected
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedValue = function (numericValue, negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = numericValue.toString();
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.negated = negated === 1;
-            value.value = numericValue;
-            this.expected = value;
-        };
-        /**
-         * This function reports an actual reference value.
-         *
-         * @param {number} referencePointer - The actual reference pointer.
-         * @param {number} offset - The size of the reference in bytes.
-         */
-        TestCollector.prototype.reportActualReference = function (referencePointer, offset) {
-            if (!this.ready) {
-                this.reportInvalidExpectCall();
-                return;
-            }
-            var value = new ActualValue_1.ActualValue();
-            value.message = "Reference Value";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.pointer = referencePointer;
-            value.offset = offset;
-            value.bytes = Array.from(this.wasm.U8.slice(referencePointer, referencePointer + offset));
-            value.value = referencePointer;
-            this.actual = value;
-        };
-        /**
-         * This function reports an expected reference value.
-         *
-         * @param {number} referencePointer - The expected reference pointer.
-         * @param {number} offset - The size of the reference in bytes.
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedReference = function (referencePointer, offset, negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = "Reference Value";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.pointer = referencePointer;
-            value.offset = offset;
-            value.bytes = Array.from(this.wasm.U8.slice(referencePointer, referencePointer + offset));
-            value.negated = negated === 1;
-            value.value = referencePointer;
-            this.expected = value;
-        };
-        /**
-         * This function reports an expected truthy value.
-         *
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedTruthy = function (negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = "Truthy Value";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.negated = negated === 1;
-            this.expected = value;
-        };
-        /**
-         * This function reports an expected falsy value.
-         *
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedFalsy = function (negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = "Falsy Value";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.negated = negated === 1;
-            this.expected = value;
-        };
-        /**
-         * This function reports an expected finite value.
-         *
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedFinite = function (negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = "Finite Value";
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.negated = negated === 1;
-            this.expected = value;
-        };
-        /**
-         * This function reports an actual string value.
-         *
-         * @param {number} stringPointer - A pointer that points to the actual string.
-         */
-        TestCollector.prototype.reportActualString = function (stringPointer) {
-            if (!this.ready) {
-                this.reportInvalidExpectCall();
-                return;
-            }
-            var value = new ActualValue_1.ActualValue();
-            value.message = this.wasm.__getString(stringPointer);
-            value.pointer = stringPointer;
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.value = stringPointer;
-            this.actual = value;
-        };
-        /**
-         * This function reports an expected string value.
-         *
-         * @param {number} stringPointer - A pointer that points to the expected string.
-         * @param {1 | 0} negated - An indicator if the expectation is negated.
-         */
-        TestCollector.prototype.reportExpectedString = function (stringPointer, negated) {
-            var value = new ActualValue_1.ActualValue();
-            value.message = this.wasm.__getString(stringPointer);
-            value.pointer = stringPointer;
-            value.stack = this.getLogStackTrace();
-            value.target = this.logTarget;
-            value.negated = negated === 1;
-            value.value = stringPointer;
-            this.expected = value;
-        };
-        /**
-         * This function overrides the provided AssemblyScript `env.abort()` function to catch abort
-         * reasons.
-         *
-         * @param {number} reasonPointer - This points to the message value that causes the expectation to
-         * fail.
-         * @param {number} _fileNamePointer - The file name that reported the error. (Ignored)
-         * @param {number} _line - The line that reported the error. (Ignored)
-         * @param {number} _col - The column that reported the error. (Ignored)
-         */
-        TestCollector.prototype.abort = function (reasonPointer, _fileNamePointer, _line, _col) {
-            this.message = this.wasm.__getString(reasonPointer);
-        };
-        /**
-         * This web assembly linked function modifies the state machine to enable
-         * performance for the following test.
-         *
-         * @param {1 | 0} value - A value indicating if performance should be enabled.
-         */
-        TestCollector.prototype.performanceEnabled = function (value) {
-            this.performanceEnabledValue = value === 1;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to set the maximum number of
-         * samples for the following test.
-         *
-         * @param {number} value - The maximum number of samples to collect for the following test.
-         */
-        TestCollector.prototype.maxSamples = function (value) {
-            this.maxSamplesValue = value;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to set the maximum amount of
-         * time to run the following test in milliseconds
-         *
-         * @param {number} value - The maximum number of milliseconds to run the following test.
-         */
-        TestCollector.prototype.maxTestRunTime = function (value) {
-            this.maxTestRunTimeValue = value;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to set the number of decimal places
-         * to round all the statistics to.
-         *
-         * @param {number} value - The number of decimal places to round to.
-         */
-        TestCollector.prototype.roundDecimalPlaces = function (value) {
-            this.roundDecimalPlacesValue = value;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to cause the next test to report
-         * an average run time.
-         *
-         * @param {1 | 0} value - A boolean indicating if the average should be reported.
-         */
-        TestCollector.prototype.reportAverage = function (value) {
-            this.recordAverageValue = value === 1;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to cause the next test to report
-         * an median run time.
-         *
-         * @param {1 | 0} value - A boolean indicating if the median should be reported.
-         */
-        TestCollector.prototype.reportMedian = function (value) {
-            this.recordMedianValue = value === 1;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to cause the next test to report
-         * a standard deviation calculation on the run times.
-         *
-         * @param {1 | 0} value - A boolean indicating if the standard deviation should be reported.
-         */
-        TestCollector.prototype.reportStdDev = function (value) {
-            this.recordStdDevValue = value === 1;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to cause the next test to report
-         * the maximum run time for this test.
-         *
-         * @param {1 | 0} value - A boolean indicating if the max should be reported.
-         */
-        TestCollector.prototype.reportMax = function (value) {
-            this.recordMaxValue = value === 1;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to cause the next test to report
-         * the minimum run time for this test.
-         *
-         * @param {1 | 0} value - A boolean indicating if the min should be reported.
-         */
-        TestCollector.prototype.reportMin = function (value) {
-            this.recordMinValue = value === 1;
-        };
-        /**
-         * This web assembly linked function modifies the state machine to cause the next test to report
-         * the variance of the run times for this test.
-         *
-         * @param {1 | 0} value - A boolean indicating if the min should be reported.
-         */
-        TestCollector.prototype.reportVariance = function (value) {
-            this.recordVariance = value === 1;
-        };
-        /**
-         * This method reports to the TestContext that an expect function call was used outside of the
-         * intended test functions.
-         */
-        TestCollector.prototype.reportInvalidExpectCall = function () {
-            this.errors.push({
-                type: "InvalidExpectCall",
-                message: "An expect() function call was used outside of a test function in " + this.fileName + ".",
-                stackTrace: this.getLogStackTrace(),
-            });
-        };
-        /**
-         * Gets an error stack trace.
-         */
-        TestCollector.prototype.getErrorStackTrace = function (ex) {
-            var stackItems = ex.stack.toString().split("\n");
-            return [stackItems[0]].concat(stackItems.slice(1).filter(wasmFilter)).join("\n");
-        };
-        /**
-         * Reset all the performance values to the configured values.
-         */
-        TestCollector.prototype.resetPerformanceValues = function () {
-            this.performanceEnabledValue = this.performanceConfiguration.enabled;
-            this.maxSamplesValue = this.performanceConfiguration.maxSamples;
-            this.maxTestRunTimeValue = this.performanceConfiguration.maxTestRunTime;
-            this.roundDecimalPlacesValue = this.performanceConfiguration.roundDecimalPlaces;
-            this.recordAverageValue = this.performanceConfiguration.reportAverage;
-            this.recordMedianValue = this.performanceConfiguration.reportMedian;
-            this.recordStdDevValue = this.performanceConfiguration.reportStandardDeviation;
-            this.recordMaxValue = this.performanceConfiguration.reportMax;
-            this.recordMinValue = this.performanceConfiguration.reportMin;
-            this.recordVariance = this.performanceConfiguration.reportVariance;
-        };
-        /**
-         * Gets a log stack trace.
-         */
-        TestCollector.prototype.getLogStackTrace = function () {
-            return new Error("Get stack trace.")
-                .stack
-                .toString()
-                .split("\n")
-                .slice(1)
-                .filter(wasmFilter)
-                .join("\n");
-        };
-        return TestCollector;
-    }());
-    exports.TestCollector = TestCollector;
-});
-//# sourceMappingURL=TestCollector.js.map
