+++ lib/test/TestContext.js
@@ -1,304 +1,392 @@
-var __extends = (this && this.__extends) || (function () {
-    var extendStatics = function (d, b) {
-        extendStatics = Object.setPrototypeOf ||
-            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
-            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
-        return extendStatics(d, b);
-    };
-    return function (d, b) {
-        extendStatics(d, b);
-        function __() { this.constructor = d; }
-        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-    };
-})();
 (function (factory) {
     if (typeof module === "object" && typeof module.exports === "object") {
         var v = factory(require, exports);
         if (v !== undefined) module.exports = v;
     }
     else if (typeof define === "function" && define.amd) {
-        define(["require", "exports", "../reporter/DefaultTestReporter", "perf_hooks", "../util/timeDifference", "./TestCollector"], factory);
+        define(["require", "exports", "../reporter/EmptyReporter", "../util/ActualValue", "../util/LogValue"], factory);
     }
 })(function (require, exports) {
     "use strict";
     Object.defineProperty(exports, "__esModule", { value: true });
-    var DefaultTestReporter_1 = require("../reporter/DefaultTestReporter");
-    var perf_hooks_1 = require("perf_hooks");
-    var timeDifference_1 = require("../util/timeDifference");
-    var TestCollector_1 = require("./TestCollector");
-    var TestContext = /** @class */ (function (_super) {
-        __extends(TestContext, _super);
+    var EmptyReporter_1 = require("../reporter/EmptyReporter");
+    var ActualValue_1 = require("../util/ActualValue");
+    var LogValue_1 = require("../util/LogValue");
+    var wasmFilter = function (input) { return /wasm-function/i.test(input); };
+    var TestContext = /** @class */ (function () {
         function TestContext(props) {
-            var _this = _super.call(this, props) || this;
-            _this.time = 0;
-            _this.pass = true;
-            _this.startupTime = 0;
-            _this.reporter = new DefaultTestReporter_1.DefaultTestReporter();
+            this.reporter = new EmptyReporter_1.EmptyReporter();
             /* istanbul ignore next */
-            _this.stdout = process ? process.stdout : null;
+            this.stdout = process ? process.stdout : null;
             /* istanbul ignore next */
-            _this.stderr = process.stderr ? process.stderr : null;
-            _this.endGroup = false;
+            this.stderr = process.stderr ? process.stderr : null;
+            this.performanceConfiguration = {};
+            this.testRegex = new RegExp("");
+            this.groupRegex = new RegExp("");
+            this.fileName = "";
+            this.wasm = null;
+            this.actual = null;
+            this.expected = null;
+            this.logTarget = {
+                logs: [],
+                stack: "",
+                message: "",
+            };
             if (props) {
-                /* istanbul ignore next */
                 if (props.reporter)
-                    _this.reporter = props.reporter;
-                /* istanbul ignore next */
+                    this.reporter = props.reporter;
                 if (props.stdout)
-                    _this.stdout = props.stdout;
-                /* istanbul ignore next */
+                    this.stdout = props.stdout;
                 if (props.stderr)
-                    _this.stderr = props.stderr;
+                    this.stderr = props.stderr;
+                if (props.performanceConfiguration)
+                    this.performanceConfiguration = props.performanceConfiguration;
+                if (props.testRegex)
+                    this.testRegex = props.testRegex;
+                if (props.groupRegex)
+                    this.groupRegex = props.groupRegex;
+                if (props.fileName)
+                    this.fileName = props.fileName;
             }
-            return _this;
         }
         /**
-         * Run the tests on the wasm module.
+         * This method creates a WebAssembly imports object with all the TestContext functions
+         * bound to the TestContext.
+         *
+         * @param {any[]} imports - Every import item specified.
          */
-        TestContext.prototype.run = function (wasm) {
-            // set wasm immediately
-            this.wasm = wasm;
-            // start the timer
-            var start = perf_hooks_1.performance.now();
-            // start the module up
-            _super.prototype.collectTests.call(this);
-            // calculate startuptime
-            this.startupTime = timeDifference_1.timeDifference(perf_hooks_1.performance.now(), start);
-            if (this.errors.length > 0)
-                return;
-            this.ready = true;
-            // start the test suite
-            this.reporter.onStart(this);
-            for (var _i = 0, _a = this.testGroups; _i < _a.length; _i++) {
-                var group = _a[_i];
-                this.runGroup(group);
+        TestContext.prototype.createImports = function () {
+            var _this = this;
+            var imports = [];
+            for (var _i = 0; _i < arguments.length; _i++) {
+                imports[_i] = arguments[_i];
             }
-            var end = perf_hooks_1.performance.now();
-            this.time = timeDifference_1.timeDifference(end, start);
-            this.reporter.onFinish(this);
-        };
-        TestContext.prototype.runGroup = function (group) {
-            this.endGroup = false;
-            // set the group starttime
-            group.start = perf_hooks_1.performance.now();
-            // set the log target
-            this.logTarget = group;
-            // for each beforeAllCallback
-            this.runBeforeAll(group);
-            // report the group as started, and log all the beforeAll logs outside the describe block
-            this.reporter.onGroupStart(group);
-            if (this.endGroup)
-                return;
-            for (var _i = 0, _a = group.tests; _i < _a.length; _i++) {
-                var result = _a[_i];
-                this.runTest(group, result);
-                if (this.endGroup)
-                    return;
-                this.reporter.onTestFinish(group, result);
-                this.logTarget = group;
-            }
-            // for each afterAllCallback
-            this.runAfterAll(group);
-            if (this.endGroup)
-                return;
-            // finish the group
-            group.end = perf_hooks_1.performance.now();
-            group.time = timeDifference_1.timeDifference(group.end, group.start);
-            group.reason = "Test suite " + group.name + " passed successfully.";
-            this.reporter.onGroupFinish(group);
+            var result = Object.assign.apply(Object, [{}].concat(imports, [{
+                    __aspect: {
+                        clearExpected: this.clearExpected.bind(this),
+                        debug: this.debug.bind(this),
+                        tryCall: this.tryCall.bind(this),
+                        logNull: this.logNull.bind(this),
+                        logReference: this.logReference.bind(this),
+                        logString: this.logString.bind(this),
+                        logValue: this.logValue.bind(this),
+                        reportActualNull: this.reportActualNull.bind(this),
+                        reportExpectedNull: this.reportExpectedNull.bind(this),
+                        reportActualValue: this.reportActualValue.bind(this),
+                        reportExpectedValue: this.reportExpectedValue.bind(this),
+                        reportActualReference: this.reportActualReference.bind(this),
+                        reportExpectedReference: this.reportExpectedReference.bind(this),
+                        reportActualString: this.reportActualString.bind(this),
+                        reportExpectedString: this.reportExpectedString.bind(this),
+                        reportExpectedTruthy: this.reportExpectedTruthy.bind(this),
+                        reportExpectedFalsy: this.reportExpectedFalsy.bind(this),
+                        reportExpectedFinite: this.reportExpectedFinite.bind(this),
+                    },
+                }]));
+            result.env = result.env || {};
+            var previousAbort = (result.env.abort) || (function () { });
+            result.env.abort = function () {
+                var args = [];
+                for (var _i = 0; _i < arguments.length; _i++) {
+                    args[_i] = arguments[_i];
+                }
+                previousAbort.apply(void 0, args);
+                // @ts-ignore
+                _this.abort.apply(_this, args);
+            };
+            return result;
         };
         /**
-         * Run a given test.
-         *
-         * @param {RunContext} runContext - The current run context.
-         * @param {TestGroup} group - The current run group.
-         * @param {number} testIndex - The current test index.
+         * This function reports an actual null value.
          */
-        TestContext.prototype.runTest = function (group, result) {
-            // set the log target
-            this.logTarget = result;
-            this.reporter.onTestStart(group, result);
-            result.ran = true;
-            result.start = perf_hooks_1.performance.now();
-            // If performance is enabled, use the performance values, otherwise, just run once.
-            if (result.performance) {
-                var runCount = 0;
-                var testStartTime = perf_hooks_1.performance.now();
-                var currentTestRunTime = 0;
-                // run the test loop
-                while (true) { // always run at least once
-                    this.runBeforeEach(group, result);
-                    /**
-                     * Especially because the performance functions are run repeatedly, if an error occurs, assume the
-                     * worst and skip the test group. These functions definitely are assumed to be safe by the test context.
-                     */
-                    if (this.endGroup)
-                        return;
-                    this.runTestCall(group, result);
-                    this.runAfterEach(group, result);
-                    if (this.endGroup)
-                        return; // check to see if the afterEach functions errored (see above)
-                    currentTestRunTime = perf_hooks_1.performance.now() - testStartTime; // calculate how long the current test has run
-                    runCount += 1; // increase the run count
-                    if (runCount >= result.maxSamples)
-                        break; // if we have reached the max sample count
-                    if (currentTestRunTime >= result.maxRuntime)
-                        break; // weve collected enough samples and the test is over
-                }
-                if (result.calculateAverageValue)
-                    result.calculateAverage();
-                if (result.calculateMaxValue)
-                    result.calculateMax();
-                if (result.calculateMedianValue)
-                    result.calculateMedian();
-                if (result.calculateMinValue)
-                    result.calculateMin();
-                if (result.calculateVarianceValue)
-                    result.calculateVariance();
-                if (result.calculateStandardDeviationValue)
-                    result.calculateStandardDeviation();
-            }
-            else {
-                this.runBeforeEach(group, result);
-                if (this.endGroup)
-                    return;
-                this.runTestCall(group, result);
-                this.runAfterEach(group, result);
-                if (this.endGroup)
-                    return;
-            }
-            result.end = perf_hooks_1.performance.now();
-            result.runTime = timeDifference_1.timeDifference(result.end, result.start);
+        TestContext.prototype.reportActualNull = function () {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "null";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.value = null;
+            this.actual = value;
         };
         /**
-         * Run the current test once and collect statistics.
+         * This function reports an expected null value.
          *
-         * @param {RunContext} runContext - The current run context.
-         * @param {TestGroup} group - The current test group.
-         * @param {TestResult} result - The current test result.
-         * @param {number} testIndex - The current test index.
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
          */
-        TestContext.prototype.runTestCall = function (group, result) {
-            var start = perf_hooks_1.performance.now();
-            var testCallResult = this.tryCall(result.functionPointer);
-            var end = perf_hooks_1.performance.now();
-            result.times.push(timeDifference_1.timeDifference(end, start));
-            result.pass = result.negated
-                ? (testCallResult === 0)
-                : (testCallResult === 1);
-            if (!result.pass) {
-                group.pass = false;
-                // if it's not negated then set the message, the actual, expected, and stack values
-                if (!result.negated) {
-                    result.message = this.message;
-                    result.actual = this.actual;
-                    result.expected = this.expected;
-                    result.stack = this.stack;
-                }
-            }
+        TestContext.prototype.reportExpectedNull = function (negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "null";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.negated = negated === 1;
+            value.value = null;
+            this.expected = value;
         };
         /**
-         * Run the afterEach callbacks before running the test.
+         * This function reports an actual numeric value.
          *
-         * @param {RunContext} runContext - The current run context.
-         * @param {TestGroup} group - The current test group.
-         * @param {TestResult} result - The current test result.
+         * @param {number} numericValue - The value to be expected.
          */
-        TestContext.prototype.runAfterEach = function (group, result) {
-            // for each afterEach callback function pointer
-            for (var _i = 0, _a = group.afterEachPointers; _i < _a.length; _i++) {
-                var afterEachCallback = _a[_i];
-                var afterEachResult = this.tryCall(afterEachCallback);
-                // if afterEach fails
-                if (afterEachResult === 0) {
-                    this.endGroup = true;
-                    group.end = result.end = perf_hooks_1.performance.now();
-                    group.pass = false;
-                    this.pass = false;
-                    group.reason = "Test suite " + group.name + " failed in afterEach callback.";
-                    result.pass = false;
-                    group.time = timeDifference_1.timeDifference(group.end, group.start);
-                    this.reporter.onTestFinish(group, result);
-                    this.reporter.onGroupFinish(group);
-                    return;
-                }
-            }
+        TestContext.prototype.reportActualValue = function (numericValue) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = numericValue.toString();
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.value = numericValue;
+            this.actual = value;
         };
         /**
-         * Run the beforeEach callbacks before running the test.
+         * This function reports an expected numeric value.
          *
-         * @param {RunContext} runContext - The current run context.
-         * @param {TestGroup} group - The current test group.
-         * @param {TestResult} result - The current test result.
+         * @param {number} numericValue - The value to be expected
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
          */
-        TestContext.prototype.runBeforeEach = function (group, result) {
-            // for each beforeEach callback function pointer
-            for (var _i = 0, _a = group.beforeEachPointers; _i < _a.length; _i++) {
-                var beforeEachCallback = _a[_i];
-                var beforeEachResult = this.tryCall(beforeEachCallback);
-                // if beforeEach fails
-                if (beforeEachResult === 0) {
-                    result.end = group.end = perf_hooks_1.performance.now();
-                    group.pass = false;
-                    group.reason = group.reason = "Test suite " + group.name + " failed in beforeEach callback.";
-                    result.pass = false;
-                    group.time = timeDifference_1.timeDifference(group.end, group.start);
-                    this.reporter.onTestFinish(group, result);
-                    this.reporter.onGroupFinish(group);
-                    this.endGroup = true;
-                    return;
-                }
-            }
+        TestContext.prototype.reportExpectedValue = function (numericValue, negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = numericValue.toString();
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.negated = negated === 1;
+            value.value = numericValue;
+            this.expected = value;
         };
         /**
-         * Run the afterAll callbacks with the given runContext and group.
+         * This function reports an actual reference value.
          *
-         * @param {RunContext} runContext - The current run context.
-         * @param {TestGroup} group - The current test group.
+         * @param {number} referencePointer - The actual reference pointer.
+         * @param {number} offset - The size of the reference in bytes.
          */
-        TestContext.prototype.runAfterAll = function (group) {
-            for (var _i = 0, _a = group.afterAllPointers; _i < _a.length; _i++) {
-                var afterAllCallback = _a[_i];
-                // call each afterAll callback
-                var afterAllResult = this.tryCall(afterAllCallback);
-                // if the test fails
-                if (afterAllResult === 0) {
-                    group.end = perf_hooks_1.performance.now();
-                    group.pass = false;
-                    group.reason = "Test suite " + group.name + " failed in afterAll callback.";
-                    this.pass = false;
-                    group.time = timeDifference_1.timeDifference(group.end, group.start);
-                    this.reporter.onGroupFinish(group);
-                    this.endGroup = true;
-                    return;
-                }
-            }
+        TestContext.prototype.reportActualReference = function (referencePointer, offset) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "Reference Value";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.pointer = referencePointer;
+            value.offset = offset;
+            value.bytes = Array.from(this.wasm.U8.slice(referencePointer, referencePointer + offset));
+            value.value = referencePointer;
+            this.actual = value;
         };
         /**
-         * Run the beforeAll callbacks with the given runContext and group.
+         * This function reports an expected reference value.
          *
-         * @param {RunContext} runContext - The current run context.
-         * @param {TestGroup} group - The current test group.
+         * @param {number} referencePointer - The expected reference pointer.
+         * @param {number} offset - The size of the reference in bytes.
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
          */
-        TestContext.prototype.runBeforeAll = function (group) {
-            for (var _i = 0, _a = group.beforeAllPointers; _i < _a.length; _i++) {
-                var beforeAllCallback = _a[_i];
-                // call each beforeAll callback
-                var beforeAllResult = this.tryCall(beforeAllCallback);
-                // if the test fails
-                if (beforeAllResult === 0) {
-                    group.end = perf_hooks_1.performance.now();
-                    group.pass = false;
-                    group.reason = "Test suite " + group.name + " failed in beforeAll callback.";
-                    this.pass = false;
-                    group.time = timeDifference_1.timeDifference(group.end, group.start);
-                    this.endGroup = true;
-                    return;
-                }
+        TestContext.prototype.reportExpectedReference = function (referencePointer, offset, negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "Reference Value";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.pointer = referencePointer;
+            value.offset = offset;
+            value.bytes = Array.from(this.wasm.U8.slice(referencePointer, referencePointer + offset));
+            value.negated = negated === 1;
+            value.value = referencePointer;
+            this.expected = value;
+        };
+        /**
+         * This function reports an expected truthy value.
+         *
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
+         */
+        TestContext.prototype.reportExpectedTruthy = function (negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "Truthy Value";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.negated = negated === 1;
+            this.expected = value;
+        };
+        /**
+         * This function reports an expected falsy value.
+         *
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
+         */
+        TestContext.prototype.reportExpectedFalsy = function (negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "Falsy Value";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.negated = negated === 1;
+            this.expected = value;
+        };
+        /**
+         * This function reports an expected finite value.
+         *
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
+         */
+        TestContext.prototype.reportExpectedFinite = function (negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = "Finite Value";
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.negated = negated === 1;
+            this.expected = value;
+        };
+        /**
+         * This function reports an actual string value.
+         *
+         * @param {number} stringPointer - A pointer that points to the actual string.
+         */
+        TestContext.prototype.reportActualString = function (stringPointer) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = this.wasm.__getString(stringPointer);
+            value.pointer = stringPointer;
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.value = stringPointer;
+            this.actual = value;
+        };
+        /**
+         * This function reports an expected string value.
+         *
+         * @param {number} stringPointer - A pointer that points to the expected string.
+         * @param {1 | 0} negated - An indicator if the expectation is negated.
+         */
+        TestContext.prototype.reportExpectedString = function (stringPointer, negated) {
+            var value = new ActualValue_1.ActualValue();
+            value.message = this.wasm.__getString(stringPointer);
+            value.pointer = stringPointer;
+            value.stack = this.getLogStackTrace();
+            value.target = this.logTarget;
+            value.negated = negated === 1;
+            value.value = stringPointer;
+            this.expected = value;
+        };
+        /**
+         * This function overrides the provided AssemblyScript `env.abort()` function to catch abort
+         * reasons.
+         *
+         * @param {number} reasonPointer - This points to the message value that causes the expectation to
+         * fail.
+         * @param {number} _fileNamePointer - The file name that reported the error. (Ignored)
+         * @param {number} _line - The line that reported the error. (Ignored)
+         * @param {number} _col - The column that reported the error. (Ignored)
+         */
+        TestContext.prototype.abort = function (reasonPointer, _fileNamePointer, _line, _col) {
+            this.logTarget.message = this.wasm.__getString(reasonPointer);
+        };
+        /**
+         * Gets a log stack trace.
+         */
+        TestContext.prototype.getLogStackTrace = function () {
+            return new Error("Get stack trace.")
+                .stack
+                .toString()
+                .split("\n")
+                .slice(1)
+                .filter(wasmFilter)
+                .join("\n");
+        };
+        /**
+         * Gets an error stack trace.
+         */
+        TestContext.prototype.getErrorStackTrace = function (ex) {
+            var stackItems = ex.stack.toString().split("\n");
+            return [stackItems[0]].concat(stackItems.slice(1).filter(wasmFilter)).join("\n");
+        };
+        /**
+         * This is called to stop the debugger.  e.g. `node --inspect-brk asp`.
+         */
+        TestContext.prototype.debug = function () { debugger; };
+        /**
+         * This is a web assembly utility function that wraps a function call in a try catch block to
+         * report success or failure.
+         *
+         * @param {number} pointer - The function pointer to call. It must accept no parameters and return
+         * void.
+         * @returns {1 | 0} - If the callback was run successfully without error, it returns 1, else it
+         * returns 0.
+         */
+        TestContext.prototype.tryCall = function (pointer) {
+            if (pointer === -1)
+                return 1;
+            try {
+                this.wasm.__call(pointer);
             }
+            catch (ex) {
+                this.logTarget.stack = this.getErrorStackTrace(ex);
+                return 0;
+            }
+            return 1;
+        };
+        /**
+         * Log a null value to the reporter.
+         */
+        TestContext.prototype.logNull = function () {
+            // create a new log value
+            var value = new LogValue_1.LogValue();
+            var target = this.logTarget;
+            // collect log metadata
+            value.stack = this.getLogStackTrace();
+            value.message = "null";
+            value.target = target;
+            // push the log value to the logs
+            target.logs.push(value);
+        };
+        /**
+         * This function is called after each expectation if the expectation passes. This prevents other
+         * unreachable() conditions that throw errors to report actual and expected values too.
+         */
+        TestContext.prototype.clearExpected = function () {
+            this.expected = null;
+            this.actual = null;
+            this.logTarget.stack = "";
+        };
+        /**
+         * Log a reference to the reporter.
+         *
+         * @param {number} referencePointer - The pointer to the reference.
+         * @param {number} offset - The offset of the reference.
+         */
+        TestContext.prototype.logReference = function (referencePointer, offset) {
+            var value = new LogValue_1.LogValue();
+            var target = this.logTarget;
+            value.bytes = Array.from(this.wasm.U8.slice(referencePointer, referencePointer + offset));
+            value.message = "Reference Type";
+            value.offset = offset;
+            value.pointer = referencePointer;
+            value.stack = this.getLogStackTrace();
+            value.target = target;
+            value.value = referencePointer;
+            // push the log value to the logs
+            target.logs.push(value);
+        };
+        /**
+         * This adds a logged string to the current test.
+         *
+         * @param {number} pointer - The pointer to the logged string reference.
+         */
+        TestContext.prototype.logString = function (pointer) {
+            var value = new LogValue_1.LogValue();
+            var target = this.logTarget;
+            value.message = this.wasm.__getString(pointer);
+            value.offset = 0;
+            value.pointer = pointer;
+            value.stack = this.getLogStackTrace();
+            value.target = target;
+            value.value = pointer;
+            // push the log value to the logs
+            target.logs.push(value);
+        };
+        /**
+         * Log a numevalueric value to the reporter.
+         *
+         * @param {number} value - The value to be logged.
+         */
+        TestContext.prototype.logValue = function (numericValue) {
+            var value = new LogValue_1.LogValue();
+            var target = this.logTarget;
+            value.stack = this.getLogStackTrace();
+            value.message = "Value " + numericValue.toString();
+            value.value = numericValue;
+            value.target = target;
+            // push the log value to the logs
+            target.logs.push(value);
         };
         return TestContext;
-    }(TestCollector_1.TestCollector));
+    }());
     exports.TestContext = TestContext;
 });
 //# sourceMappingURL=TestContext.js.map
