+++ lib/util/ILogTarget.d.ts
@@ -1,5 +1,7 @@
 import { LogValue } from "./LogValue";
 export interface ILogTarget {
+    message: string;
     logs: LogValue[];
+    stack: string;
 }
 //# sourceMappingURL=ILogTarget.d.ts.map
