--- src/test/IWarning.ts
@@ -1,5 +1,0 @@
-export interface IWarning {
-  type: string;
-  message: string;
-  stackTrace: string;
-}
