+++ src/util/IAspectExports.ts
@@ -1,4 +1,14 @@
 export interface IAspectExports {
   __call(pointer: number): void;
-  __main(): void;
+  __run(): void;
+  __set_performanceEnabled(value: 1 | 0): void;
+  __set_maxSamples(value: number): void;
+  __set_maxTestRunTime(value: number): void;
+  __set_roundDecimalPlaces(value: number): void;
+  __set_recordAverage(value: 1 | 0): void;
+  __set_recordMedian(value: 1 | 0): void;
+  __set_recordStdDev(value: 1 | 0): void;
+  __set_recordMax(value: 1 | 0): void;
+  __set_recordMin(value: 1 | 0): void;
+  __set_recordVar(value: 1 | 0): void;
 }
